% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyUpsimpler.R
\name{applyUpsimpler}
\alias{applyUpsimpler}
\title{Apply upsimpler on a selection of target samples}
\usage{
applyUpsimpler(
  upsimpler,
  upsimplerArgs,
  metadataDF,
  targetSampleIDs,
  classColumn,
  sampleColumn
)
}
\arguments{
\item{upsimpler}{The (already initialized) upsimpler object to be used for generating the synthetic samples.}

\item{upsimplerArgs}{List of arguments to be passed to the upsimpler algorithm.}

\item{metadataDF}{Dataframe containing the metadata of the samples (samples as rows).}

\item{targetSampleIDs}{List of sample IDs to be used as seed for the upsimpler algorithm.
Must be present in the dataDF and metadataDF in upsimplerArgs.}

\item{classColumn}{Column name in the metadata dataframe containing the class information.}

\item{sampleColumn}{Column name in the metadata dataframe containing the sample identifier.}
}
\value{
Dataframe containing the synthetic samples generated by the upsimpler algorithm
for all target samples (seeds) combined ($synthDataDF), as well as the metadata of the synthetic samples ($synthMetadataDF).
The metadata contain only the class information of the synthetic samples.
}
\description{
Function to apply the upsimpler algorithm on a single dataset.
Given a list of target samples, this function generates the requested number of synthetic samples
per target sample (seed) and returns them as a single dataframe.
}
