% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeanAndSDOverall.R
\name{calculateMeanAndSDOverall}
\alias{calculateMeanAndSDOverall}
\title{Calculate mean and standard deviation for whole dataset}
\usage{
calculateMeanAndSDOverall(
  minorityDir,
  majorityDir,
  metaDataTest,
  subtype = F,
  subset = F,
  probabilityThreshold
)
}
\arguments{
\item{minorityDir}{Directory in which the minority model(s) are stored.}

\item{majorityDir}{Directory in which the majority model(s) are stored.}

\item{metaDataTest}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the test set.}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?}

\item{subset}{Do you want to only include a subset of the data for the performance measure calculations? If so, add their rownames here.}

\item{probabilityThreshold}{What is the threshold you would like to use to call a classification 'confident'?}

\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}

\item{crossValidation}{Specify whether the results are from the cross-validation setup or not.}
}
\value{
Dataframe with performance for all samples within the dataset combined,
instead of with stratification by tumor (sub)type frequency within reference cohort.
}
\description{
Calculate mean and standard deviation for whole dataset
}
