% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMulticlassROC.R
\name{calculateMulticlassROC}
\alias{calculateMulticlassROC}
\title{Title}
\usage{
calculateMulticlassROC(
  predictionsMMFinal,
  metaDataRef,
  classColumn,
  MMProbabilityList,
  filtering,
  probabilityThreshold
)
}
\arguments{
\item{predictionsMMFinal}{Dataframe containing the final classification ($predict) and original diagnosis label ($originalCall).}

\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{MMProbabilityList}{List with the probability scores for all classification labels that were assigned to samples.}

\item{filtering}{Do you want to use filtered classifications only?}

\item{probabilityThreshold}{What is the threshold you would like to use to call a classification 'confident'?}
}
\value{
No idea.
}
\description{
Title
}
