% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeSubtypeNameToType.R
\name{changeSubtypeNameToType}
\alias{changeSubtypeNameToType}
\title{Link tumor subtype to tumor type}
\usage{
changeSubtypeNameToType(
  probabilities,
  linkClassAndHigherClass,
  classColumn,
  higherClassColumn
)
}
\arguments{
\item{probabilities}{List with the probabilities from the different tumor subtype predictions from the minority or majority classifier.}

\item{linkClassAndHigherClass}{dataframe containing the link between the tumor subtypes and the tumor types.
The dataframe is automatically generated in previous functions from the metadata.}
}
\value{
List of all the samples containing the probabilities for the different tumor type predictions from the minority or majority classifier.
}
\description{
Function to find the associated tumor type for each tumor subtype prediction.
}
