% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Clean reference cohort}
\usage{
cleanData(
  refCohort,
  classColumn = "Disease_sub_specification1",
  n = 3,
  inputDir =
    "~/surfdrive/Shared/Kemmeren group/Research_Projects/RNA_classification_FW/data/input/CSR/2023_03/"
)
}
\arguments{
\item{refCohort}{R-object containing both the RNA-seq count data matrix ($counts) and the associated metadata ($metaData).}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{n}{What is the minimum number of samples that is required per tumor (sub)type?}

\item{inputDir}{Directory where the CSR-data is stored that is needed to check whether
there is only one data entry per patient per tumor type.}
}
\value{
R-object containing the cleaned RNA-seq count data matrix ($counts) and the associated metadata ($metaData).
}
\description{
Function to remove duplicate samples from the same patient from the reference cohort,
and tumor (sub)types for which there are not enough samples available (< \emph{n}).
}
