\name{coef.mvl1lm}
\Rdversion{1.1}
\alias{coef.mvl1lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coefficients of an mvl1lm Object
}
\description{
Extracts the coefficents of an mvl1lm object.
}
\usage{
\method{coef}{mvl1lm}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
an object of class \code{mvl1lm}.
}
  \item{\dots}{
needed for other methods.
}
}
\details{
Note that for \code{rank} scores the intercept, even when specified in the model, is not considered a coefficient.
}
\author{Klaus Nordhausen, \email{klaus.nordhausen@uta.fi} }

\keyword{methods}


