% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSeedPredictions.R
\name{combineSeedPredictions}
\alias{combineSeedPredictions}
\title{Combine the results of different cross-validation runs}
\usage{
combineSeedPredictions(
  minorityDir,
  majorityDir,
  subtype = F,
  higherClassColumn,
  classColumn
)
}
\arguments{
\item{minorityDir}{Directory in which the minority model(s) are stored.}

\item{majorityDir}{Directory in which the majority model(s) are stored.}

\item{subtype}{Do you want to combine the classifications on the subtype level?}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}

\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}
}
\value{
List containing the a dataframe with the final predictions (predictionsMMFinal),
and a list with the probability scores for all classification labels that were assigned to samples (MMProbabilityList).

predictionsMMFinal contains the top 3 final predictions ($predict{2,3}) with their accompanying probability scores
($probability{1,2,3} and the original diagnosis label ($originalCall).
}
\description{
To check for algorithm stability, cross-validation setups can be run multiple times
to see whether results remain similar. This function combines the results of different runs into one,
coming with a final prediction for each sample used during the cross-validation setup.
}
