% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSeparateClassifierAccuracies.R
\name{combineSeparateClassifierAccuracies}
\alias{combineSeparateClassifierAccuracies}
\title{Get results for separate classifiers as compared to M&M}
\usage{
combineSeparateClassifierAccuracies(
  minorityDir,
  majorityDir,
  probabilityThreshold = 0.8,
  probabilityThresholdMajority = 0.9,
  probabilityThresholdMinority = 0.73,
  subtype,
  returnPlot = T
)
}
\arguments{
\item{minorityDir}{Directory in which the minority model(s) are stored.}

\item{majorityDir}{Directory in which the majority model(s) are stored.}

\item{probabilityThreshold}{What is the probability score threshold you would like to use to call a
classification 'confident' for a M&M prediction?}

\item{probabilityThresholdMajority}{What is the probability score threshold you would like to use to call a
classification 'confident' for a Majority Classifier prediction? Default value established based on ROC-curve analyses.}

\item{probabilityThresholdMinority}{What is the probability score threshold you would like to use to call a
classification 'confident' for a Minority Classifier prediction? Default value established based on ROC-curve analyses.}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?}

\item{returnPlot}{Do you want to obtain the data or (FALSE) or get the resulting plot (TRUE)?}
}
\value{
If returnPlot = T: Plot with the separate classifier accuracy (Minority Classifier,
Majority Classifier and M&M) stratified on population frequency.
If returnPlot = F: Dataframe containing the fraction correct labels within the top1, top2 and top3 classification labels.
}
\description{
Get results for separate classifiers as compared to M&M
}
