% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertResultToClassification.R
\name{convertResultToClassification}
\alias{convertResultToClassification}
\title{Combine individual model results into final classification}
\usage{
convertResultToClassification(
  result,
  metaDataRef,
  addOriginalCall,
  classColumn = NA
)
}
\arguments{
\item{result}{Dataframe containing the classification labels from each model for each sample.}

\item{metaDataRef}{Metadata file containing the links between the samples and the tumor domain, type and subtype diagnosis.}

\item{addOriginalCall}{Would you like to add the original tumor subtype label from the metadata (TRUE)?}

\item{classColumn}{Name of column in the metadata file that contains the tumor subtype labels.}
}
\value{
List containing the final per-sample classifications ($classifications),
per-label probabilities ($probabilityList) and metadata that the classifications were based on ($metaDataRef).
}
\description{
Convert individual classifications to an object containing the final per-sample classifications,
per-label probabilities and metadata that the classifications were based on.
}
