% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExtraData.R
\name{createExtraData}
\alias{createExtraData}
\title{Create fake count data entries}
\usage{
createExtraData(countDataRef, classesWith2)
}
\arguments{
\item{countDataRef}{Dataframe containing the RNA-transcript per million data. Patient samples are in the rows,
different genes in the columns.The data has been cleaned by the ribodepletion correction protocol.}

\item{classesWith2}{A vector containing the tumor (sub)types that currently
only have two samples within the metadata file.}
}
\value{
Count-data dataframe now including extra fake entries, with a minimum of
three entries per tumor (sub)type.
}
\description{
This function is designed to create new count data entries for
tumor (sub)types that only have two samples.
Additional entries are needed for analysis of variance (ANOVA),
as each group requires at least three entries to calculate a mean and a standard deviation for each gene.
These entries are calculated by taking the mean values for each gene from the two available samples,
giving an 'average' sample as the third input for the ANOVA.
These extra samples are only used during the gene selection procedure,
not during the classification.
}
