% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExtraMetaData.R
\name{createExtraMetaData}
\alias{createExtraMetaData}
\title{Create synthetic metadata entries}
\usage{
createExtraMetaData(metaDataRef, classesWith2)
}
\arguments{
\item{metaDataRef}{Original metadata file that contains the tumor (sub)type labels.}

\item{classesWith2}{A vector containing the tumor (sub)types that currently
only have two samples within the metadata file.}
}
\value{
Metadata file now including extra fake entries, with a minimum of
three entries per tumor (sub)type.
}
\description{
This function is designed to create new metadata entries for
tumor (sub)types that only have two samples.
Additional entries are needed for analysis of variance (ANOVA),
as each group requires at least three entries to calculate a mean and a standard deviation for each gene.
}
