% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFeatureDF.R
\name{createFeatureDF}
\alias{createFeatureDF}
\title{Extract importance scores from RF-models}
\usage{
createFeatureDF(modelList, whichAccuracyMeasure, nANOVAgenes)
}
\arguments{
\item{modelList}{List containing different weighted RF models with importance scores.}

\item{whichAccuracyMeasure}{Which statistic calculated in the importance score of the RF
models do you want to use, the mean decrease in accuracy (MeanDecreaseAccuracy)
or the mean decrease in the Gini score (MeanDecreaseGini)?}

\item{nANOVAgenes}{How many genes did we select during the ANOVA procedure?}
}
\value{
Dataframe containing the mean decreases in accuracy / Gini score from removing
different genes during a weighted RF procedure for different models.
The genes are in the rows, the different models in the columns.
}
\description{
This function is designed to extract all the mean decreases in accuracy / Gini score
from a performed weighted Random Forest (RF). The mean decreases in accuracy / Gini score
from all different RNA-transcripts selected in the ANOVA-procedure is calculated,
for all models within the modelList.
}
