% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModelsMinority.R
\name{createModelsMinority}
\alias{createModelsMinority}
\title{Generate Minority classifier models}
\usage{
createModelsMinority(
  countDataRef,
  metaDataRef,
  classColumn,
  higherClassColumn,
  domainColumn,
  sampleColumn,
  nModels = 100,
  meanExpression = 5,
  nANOVAgenes = 1000,
  maxSamplesPerType = 3,
  ntree = 500,
  nFeatures = 300,
  whichSeed = 1,
  outputDir = paste0("./", format(as.Date(Sys.Date(), "\%Y-\%m-\%d"), "\%Y_\%m_\%d")),
  proteinCodingGenes,
  saveModel = T
)
}
\arguments{
\item{countDataRef}{Matrix containing the RNA-transcript per million data. Samples are in the columns,
different RNA-transcripts in the rows.}

\item{metaDataRef}{Metadata file containing the links between the samples and the tumor domain, type and subtype diagnosis.}

\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}

\item{domainColumn}{Column in the metadata file that contains the tumor domain labels.}

\item{sampleColumn}{Column in the metadata file that contains the samples.}

\item{nModels}{How many models should be created for the Minority classifier?
Default is 100.}

\item{meanExpression}{Selection criterion for the RNA-transcripts,
specifying what the minimum mean expression of a RNA-transcripts should be for it to be included in the F-statistic analysis.
Default is 5.}

\item{nANOVAgenes}{How many RNA-transcripts should we select using the F-statistic of ANOVA? Default is 1000.}

\item{maxSamplesPerType}{How many samples should we maximally use per tumor subtype? Default is 3.}

\item{ntree}{How many trees should we use during the weighted Random Forest procedure? Default is 500.}

\item{nFeatures}{How many features should we keep after determining the most important RNA-transcripts using the Random Forest Importance Score?}

\item{whichSeed}{For reproducibility, the seed can be specified with this parameter. Default is 1.}

\item{outputDir}{Directory in which you would like to store the R-object containing the results. Default is today's date.}

\item{proteinCodingGenes}{What are the names of the RNA-transcripts that stand for protein-coding RNA-transcripts within our dataset?
Please supply it as a vector. This is needed for ribo-depletion correction model.}

\item{saveModel}{Do you want to save your generated scalings in an R-object? Default is TRUE.}
}
\value{
R-object containing the generated RF-models ($modelList), the model for the ribodepletion correction ($riboModelList),
the features that were eventually used for the weighted RF within the different folds ($reducedFeaturesList),
the metadata file associated to the reference cohort ($metaData)
and the metadata for the performed run ($metaDataRun).
}
\description{
Setup to obtain the models for the Minority classifier.
}
