% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPieChartImage.R
\name{createPieChartImage}
\alias{createPieChartImage}
\title{Plot pie-chart per domain as cohort overview}
\usage{
createPieChartImage(
  metaDataRef,
  classColumn,
  higherClassColumn,
  domainColumn = NA,
  abbreviations = NA,
  whichDomain = NA,
  storeLocation = ".",
  saveImage = F,
  textSizeClass = 0.5,
  textSizeSubspec = 0.45,
  plotColors = NA,
  includeNumbers = T
)
}
\arguments{
\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis.}

\item{classColumn}{Name of column in the metadata file that contains the tumor subtype labels.}

\item{higherClassColumn}{Name of column in the metadata file that contains the tumor type labels.}

\item{domainColumn}{Name of column in the metadata file that contains the domain labels.}

\item{abbreviations}{Dataframe containing the different tumor subtype labels with their appropriate abbreviations ($abbreviationSubtype),
and tumor type labels with their appropriate abbreviations (abbreviationTumorType).
Please make sure that the higherClassColumn name is used as column name within abbreviations for the tumor type labels,
and classColumn name for the tumor subtype labels.}

\item{whichDomain}{Which domain do you want to plot? This name should be present in the
domainColumn of the metadata.}

\item{storeLocation}{Where do you want to store your generated image? Default is within working directory.}

\item{saveImage}{Do you want to save your image? Default is FALSE.}

\item{textSizeClass}{What size should the letters have for the tumor type? Default is 0.5.}

\item{textSizeSubspec}{What size should the letters have for the tumor subtype? Default is 0.45.}

\item{plotColors}{Which colors do you want to use to color your plot? If not specified,
default colors from RColorBrewer palette spectral will be used.}

\item{includeNumbers}{Do you want to show the numbers for the pie chart on the outer edge?
Default is includeNumbers = TRUE.}
}
\value{
Pie-chart of how many samples are present within the dataset belonging to a
certain tumor type, and to a certain tumor subtype.
The pie-chart is generated for one domain, as there are many different tumor types and subtypes
within each domain.
}
\description{
Function to automatically generate a pie-chart for the samples that belong to
one of the tumor domains within the dataset. Required are dedicated dataframes
to specify the names that you would like to use for the different tumor types and subtypes,
as usually their full names are too lengthy to fit on the plot.
}
\details{
Also, the colors you want to use within the plot could be specified,
so that the sheer amount of tumor types does not cause a decreased visibility.
}
