% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createScalingsMajority.R
\name{createScalingsMajority}
\alias{createScalingsMajority}
\title{Create transformation information Majority Classifier}
\usage{
createScalingsMajority(
  countDataRef,
  metaDataRef,
  classColumn,
  patientColumn,
  nModels = 100,
  maxSamplesPerType = 50,
  nComps = 100,
  nFeatures = 2500,
  maxNeighbours = 25,
  whichSeed = 1,
  outputDir = "./",
  proteinDir,
  saveModel = T,
  saveRiboModels = F
)
}
\arguments{
\item{countDataRef}{Matrix containing the RNA-transcript per million data. Patients are in the columns,
different genes in the rows.}

\item{metaDataRef}{Metadata file containing the links between the patients and
the tumor (sub)type diagnosis within the reference cohort.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{patientColumn}{Column in the metadata file that contains the patient labels.}

\item{nModels}{How many models should be created for the majority voting system?}

\item{maxSamplesPerType}{How many samples should we maximally use per tumor (sub)type?}

\item{nComps}{How many principal components will be selected after PCA?}

\item{nFeatures}{How many of the most variable genes within the dataset should we select for principal component analysis (PCA)?}

\item{maxNeighbours}{What is the maximum number of neigbours to be used for the weighted \emph{k}-nearest neighbor algorithm?}

\item{whichSeed}{For reproducibility, the seed can be specified with this parameter.}

\item{outputDir}{Directory in which you would like to store the R-object containing the results.}

\item{proteinDir}{In which directory can we find the file specifying the names of protein-coding genes within our dataset?}
}
\value{
R-object containing the rotations and scalings
for each reference cohort subset($rotationsAndScalingList),
the model to correct for the ribodepletion efficacy ($riboModelList),
which samples were present in each subset ($samplesTrainDefList),
which genes were considered for transformation in the analysis ($nonZeroGenes),
the metadata file associated to the reference cohort ($metaData),
and metadata for the performed run ($metaDataRun).
}
\description{
This function is used to obtain the PCA-rotations and data scalings around zero
that are generate within the Majority Classifier. These rotations and scalings are needed
to transform new samples as well.
}
