% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIndividualValuesF1.R
\name{extractIndividualValuesF1}
\alias{extractIndividualValuesF1}
\title{Title}
\usage{
extractIndividualValuesF1(
  predictionsMM,
  metaDataRef,
  classColumn,
  probabilityThreshold,
  filterOrNot
)
}
\arguments{
\item{filterOrNot}{}
}
\description{
Title
}
