\name{fitted.mvl1lm}
\Rdversion{1.1}
\alias{fitted.mvl1lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fitted Values of an mvl1lm Object
}
\description{
Extracts the fitted values of an mvl1lm object.
}
\usage{
\method{fitted}{mvl1lm}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
an object of class \code{mvl1lm}.
}
  \item{\dots}{
needed for other methods.
}
}
\author{Klaus Nordhausen}

\keyword{methods}

