% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccuraciesPerTumorTypeSize.R
\name{getAccuraciesPerTumorTypeSize}
\alias{getAccuraciesPerTumorTypeSize}
\title{Obtain accuracies per tumor (sub)type sample size}
\usage{
getAccuraciesPerTumorTypeSize(
  predictionsMM,
  metaDataRef,
  classColumn,
  rounding = T,
  probabilityThreshold = 0.8
)
}
\arguments{
\item{predictionsMM}{Dataframe showing the top 3 predictions for the tumor (sub)type, together with their probability scores.}

\item{metaDataRef}{Metadata-file for the reference cohort containing the tumor (sub)type labels.}

\item{classColumn}{Which column within the metadata file contains the tumor (sub)type labels?}
}
\value{
Dataframe showing the accuracies for the different tumor size
}
\description{
Obtain accuracies per tumor (sub)type sample size
}
