% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccuracyPerClass.R
\name{getAccuracyPerClass}
\alias{getAccuracyPerClass}
\title{Obtain accuracy per tumor (sub)type}
\usage{
getAccuracyPerClass(predictions)
}
\arguments{
\item{predictions}{Dataframe containing the diagnosis label ($originalCall) and the top-ranked classification label ($predict) from a classifier.}
}
\value{
Dataframe containing the accuracy for the different tumor (sub)types within the dataset.
}
\description{
Obtain accuracy per tumor (sub)type
}
