% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnovaResults.R
\name{getAnovaResults}
\alias{getAnovaResults}
\title{Calculate F-statistic ANOVA for gene selection}
\usage{
getAnovaResults(allMaterial, allGenes, classColumn)
}
\arguments{
\item{allMaterial}{Dataframe containing both the cleaned RNA-seq count data and the metadata
containing the tumor (sub)type labels. This dataframe is a concatenation of the count-data and metadata files used before.}

\item{allGenes}{All genes that should be compared between tumor (sub)types using the ANOVA-setup.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}
}
\value{
Dataframe containing the genes with their associated F-scores and p-scores from ANOVA.
}
\description{
This function calculates the differences between the gene expression of all tumor (sub)types
from the reference cohort using an analysis of variance (ANOVA) setup, assuming equal variance.
In this way, the genes that are most variable between
the tumor (sub)types can be selected, based on their F-scores.
}
