% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfusionMatrixPlot.R
\name{getConfusionMatrixPlot}
\alias{getConfusionMatrixPlot}
\title{Get confusion matrix information}
\usage{
getConfusionMatrixPlot(
  minorityDir,
  majorityDir,
  abbreviations = NA,
  subtype,
  defineTumorWithColor = F,
  probabilityScore = 0.8
)
}
\arguments{
\item{abbreviations}{Dataframe containing the links between the tumor (sub)type,
the abbreviation required in the plot, and the domain.}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?
for being classified or non-classified?}

\item{defineTumorWithColor}{Do you want to have an extra column where the tumors can
be specified with a color?}

\item{probabilityScore}{Which probability score do you want to use as a cutoff}
}
\value{
Dataframe specifying how often certain reference-prediction
combinations are present.
}
\description{
This function is designed to extract which tiles of the confusion matrix
do contain reference-prediction information.
}
