% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCumulativeAccuracy.R
\name{getCumulativeAccuracy}
\alias{getCumulativeAccuracy}
\title{Calculate accuracy for score blocks combined}
\usage{
getCumulativeAccuracy(totalAccDF, totalSamples)
}
\arguments{
\item{totalAccDF}{Dataframe showing what is the corresponding score block ($Filter),
how many samples fall within the probability score range ($total),
how many of these samples receive a correct prediction ($totalCorrect),
how many an incorrect prediction ($totalIncorrect),
and what is the accuracy within the score block ($accuracy).}
}
\value{
The same dataframe as before, but now containing new columns $cumulativeAccuracy
(combined accuracy also with the higher score blocks),
and $cumulativeTotal (how many samples are classified in the current and higher score blocks).
}
\description{
This function calculates for all samples with a higher probability score than a certain treshold what is the accuracy.
Furthermore, the total number of samples that has a probability score higher than the threshold is determined.
}
