% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getF1ScoreMetrics.R
\name{getF1ScoreMetrics}
\alias{getF1ScoreMetrics}
\title{Check out the F1 score metrics for whole dataset}
\usage{
getF1ScoreMetrics(predictionsMM)
}
\arguments{
\item{predictionsMM}{Final predictions with the top 3 final predictions ($predict{2,3})
with their accompanying probability scores ($probability{1,2,3} and the original diagnosis label ($originalCall).}
}
\value{
Dataframe containing the macro-F1 score for the unfiltered (macroF1) and filtered dataset (macroF1FilteredWithNA, macroF1Filtered), median F1 score
for the unfiltered (medianF1) and filtered dataset (medianF1Filtered).
}
\description{
Check out the F1 score metrics for whole dataset
}
