% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMajorityPredictions.R
\name{getMajorityPredictions}
\alias{getMajorityPredictions}
\title{Obtain M&M majority predictions}
\usage{
getMajorityPredictions(
  minority,
  MMProbabilityList,
  higherClassColumn,
  crossValidation,
  metaDataRef,
  subtype
)
}
\arguments{
\item{minority}{R-object that contains the results from the Minority classifier.}

\item{MMProbabilityList}{List of all the samples containing the probabilities for the different sample predictions from the integrated M&M classifier.}

\item{higherClassColumn}{Column name within metadata-file that contains the cancer type labels.}

\item{crossValidation}{Specify whether the results are from the cross-validation setup or not.}

\item{metaDataRef}{Metadata-file for the reference cohort.}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?
If so, use \emph{subtype = T}. If you want to obtain tumor type predictions instead, use \emph{subtype = F}.}
}
\value{
Dataframe showing the top 3 predictions for the tumor (sub)type, together with their probability scores.
}
\description{
Function to extract the top 3 tumor (sub)type predictions for samples, with their accompanying probability scores.
}
