% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMMProbabilities.R
\name{getMMProbabilities}
\alias{getMMProbabilities}
\title{Integrate Minority & Majority probability scores}
\usage{
getMMProbabilities(majorityProbability, minorityProbability)
}
\arguments{
\item{majorityProbability}{List of all samples containing the probabilities for the different sample predictions from the Majority classifier.}

\item{minorityProbability}{List of all samples containing the probabilities for the different sample predictions from the Minority classifier.}
}
\value{
List of all the samples containing the probabilities for the different sample classifications from the integrated M&M classifier.
}
\description{
This function calculates the probability for each prediction,
summing up the probabilities for the Minority and Majority classifiers and dividing the probabilities either
by 2 (when both classifiers made the prediction) or 10 (when only one classifier made the prediction).
}
