% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMulticlassROCObject.R
\name{getMulticlassROCObject}
\alias{getMulticlassROCObject}
\title{Obtain probabilities per sample per class}
\usage{
getMulticlassROCObject(metaDataRef, classColumn, MMProbabilityList)
}
\arguments{
\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{MMProbabilityList}{List with the probability scores for all classification labels that were assigned to samples.}
}
\value{
Dataframe containing the probabilities per class
}
\description{
Obtain probabilities per sample per class
}
