% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPieDF.R
\name{getPieDF}
\alias{getPieDF}
\title{Convert metadata into frequency dataframe}
\usage{
getPieDF(metaDataRef, abbreviations, higherClassColumn, classColumn)
}
\arguments{
\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis.}

\item{abbreviations}{Dataframe containing the different tumor subtype labels (classColumn) with their appropriate abbreviations ($abbreviationSubtype),
and tumor type labels (higherClassColumn) with their appropriate abbreviations (abbreviationTumorType).
Please make sure that the higherClassColumn name is used as column name within abbreviationsCombi for the tumor type labels,
and classColumn name for the tumor subtype labels.}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}

\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}
}
\value{
Dataframe containing the different labels within our dataset with their frequencies.
}
\description{
Convert metadata into frequency dataframe
}
