% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrincipalComponents.R
\name{getPrincipalComponents}
\alias{getPrincipalComponents}
\title{Get PCA transformation Majority Classifier}
\usage{
getPrincipalComponents(samplesPerModel, nComps = 100)
}
\arguments{
\item{samplesPerModel}{List of different training data sample subsets used for subsetting the available training dataset.
These contain the actual samples, not just the IDs (as they were selected by \code{selectFeaturesMajority} or
as returned by \code{applyUpsimplerHomogeneously} or \code{applyUpsimplerHeterogeneously}).}

\item{nComps}{How many principal components will be selected after PCA?}
}
\value{
List containing both the derived PCA-transformation information ($prList)
and the information needed to scale new sample input data to center the features around 0 ($scaleFeaturesList).
}
\description{
Function to find the optimal principal component analysis (PCA) transformation of the reference cohort
so that a maximal variance within the data is explained. With the parameters, the eventual total number
of principal components can be adjusted.
}
