% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopClassifications.R
\name{getTopClassifications}
\alias{getTopClassifications}
\title{Obtain M&M's top classifications}
\usage{
getTopClassifications(minority, MMProbabilityList, higherClassColumn, subtype)
}
\arguments{
\item{minority}{R-object that contains the results from the Minority classifier.}

\item{MMProbabilityList}{List of all the samples containing the probabilities for the different sample predictions from the integrated M&M classifier.}

\item{higherClassColumn}{Name of column name within metadata-file that contains the tumor type labels.}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?
If so, use \emph{subtype = T}. If you want to obtain tumor type predictions instead, use \emph{subtype = F}.}
}
\value{
Dataframe showing the top 3 classifications for the tumor (sub)type ($predict{2,3}), together with their probability scores ($probability{1,2,3}.
}
\description{
Function to extract the top 3 tumor (sub)type classifications for samples, with their accompanying probability scores.
}
