% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarFeaturesMajority.R
\name{getVarFeaturesMajority}
\alias{getVarFeaturesMajority}
\title{Extract most variable features from training dataset}
\usage{
getVarFeaturesMajority(
  dataTrain,
  samplesTrainDefList,
  nFeatures = 2500,
  nModels = 100
)
}
\arguments{
\item{dataTrain}{Data to be used for model training, containing the all samples available within the training dataset.}

\item{samplesTrainDefList}{List of different training data sample subsets used for subsetting the available training dataset (dataTrain).}

\item{nFeatures}{How many of the most variable RNA-transcripts within the dataset should we select for principal component analysis (PCA)?}

\item{nModels}{How many models should be created for the classifier?}
}
\value{
List containing the most variable features, portraying information needed to scale new sample input data to center the features around 0.
}
\description{
Extract most variable features from training dataset
}
