% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howManySamplesPassThreshold.R
\name{howManySamplesPassThreshold}
\alias{howManySamplesPassThreshold}
\title{Calculate performance stratified on recall rate}
\usage{
howManySamplesPassThreshold(predictionsMM, probabilityThreshold)
}
\arguments{
\item{predictionsMM}{Dataframe containing the top 3 final classification labels ($predict{2,3}) with their accompanying probability scores
($probability{1,2,3} and the original diagnosis label ($originalCall).}

\item{probabilityThreshold}{What is the probability score threshold you would like to use to call a
classification 'confident' for a M&M prediction?}
}
\value{
Dataframe containing the accuracy of tumor entities stratified based on their recall rate.
}
\description{
Calculate performance stratified on recall rate
}
