% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceTumorHeterogeneity.R
\name{InfluenceTumorHeterogeneity}
\alias{InfluenceTumorHeterogeneity}
\title{Calculate results for single dataset concerning tumor heterogeneity}
\usage{
InfluenceTumorHeterogeneity(
  metaDataRef,
  predictionsMM,
  predictionsMMSubtype,
  trainOrTest,
  probabilityScoreTumor,
  probabilityScoreSubtype,
  metaDataFFPE = NA,
  predictionsFFPE,
  predictionsFFPESubtype
)
}
\arguments{
\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the reference cohort.}

\item{predictionsMM}{Dataframe containing the top 3 final tumor type classification labels ($predict{2,3}) with their accompanying probability scores
($probability{1,2,3} and the original diagnosis label ($originalCall).}

\item{predictionsMMSubtype}{Dataframe containing the top 3 final tumor subtype classification labels ($predict{2,3}) with their accompanying probability scores
($probability{1,2,3} and the original diagnosis label ($originalCall).}

\item{trainOrTest}{Do you want to calculate for the reference cohort ("Train") or independent test set ("Test")?}

\item{probabilityScoreTumor}{What is the probability score threshold you would like to use to call a classification 'confident' on the tumor type level?}

\item{probabilityScoreSubtype}{What is the probability score threshold you would like to use to call a classification 'confident' on the tumor subtype level?}

\item{metaDataFFPE}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis for the FFPE samples.}

\item{predictionsFFPE}{Predictions for the FFPE samples on the tumor type level.}

\item{predictionsFFPESubtype}{Predictions for the FFPE samples on the tumor subtype level.}
}
\value{
Dataframe containing the results for different sources of heterogeneity within one dataset.
}
\description{
Calculate results for single dataset concerning tumor heterogeneity
}
