% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateMM.R
\name{integrateMM}
\alias{integrateMM}
\title{Integrate results Minority & Majority classifier}
\usage{
integrateMM(
  minority,
  majority,
  metaDataRef,
  nModels,
  subtype,
  classColumn,
  higherClassColumn,
  crossValidation = T
)
}
\arguments{
\item{minority}{R-object that contains the results from the Minority classifier}

\item{majority}{R-object that contains the results from the Majority classifier}

\item{metaDataRef}{Metadata file containing the links between the patients and
the tumor (sub)type diagnosis within the reference cohort.}

\item{nModels}{How many models were used to obtain a final prediction?}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?
If so, use \emph{subtype = T}. If you want to obtain tumor type predictions instead, use \emph{subtype = F}.}

\item{classColumn}{Column name within metadata-file that contains the cancer subtype-labels.}

\item{higherClassColumn}{Column name within metadata-file that contains the cancer type labels.}

\item{crossValidation}{Specify whether the results are from the cross-validation setup or not.
This is important, as for the cross-validation setup there is a ground truth ($originalCall), while for new predictions there is not.}

\item{substituteNames}{Do you want to substitute some names?}

\item{substituteBy}{What do you want to substitute it by?}
}
\value{
Dataframe showing the top 3 predictions for the tumor (sub)type, together with their probability scores.
}
\description{
Function to obtain final predictions for the M&M algorithm by combining the results of the Minority & Majority classifiers.
}
