% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadColors.R
\name{loadColors}
\alias{loadColors}
\title{Load colors for tumor types plotting}
\usage{
loadColors()
}
\value{
Loaded R-objects containing the colors for the three domains.
}
\description{
Load colors for tumor types plotting
}
