\name{MNM-package}
\alias{MNM-package}
\alias{MNM}
\docType{package}
\title{
\packageTitle{MNM}}
\description{
Multivariate tests, estimates and methods based on the identity score, spatial sign score and spatial rank score are provided. The methods include one and c-sample problems, shape estimation and testing, 
linear regression and principal components.
}
\details{
\packageDESCRIPTION{MNM}

The methods implemented here are mainly described in Oja (2010) and the package can be used to reproduce most of the examples in the book. The book will be refered to as the MNM book. 
}
\references{
\cite{Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R. An Approach Based on Spatial Signs and Ranks}, Springer. <doi:10.1007/978-1-4419-0468-3>. } 

\cite{Nordhausen, K. and Oja, H. (2011), Multivariate L1 Methods: The Package MNM, \emph{Journal of Statistical Software}, \bold{43}, 1-28. <doi:10.18637/jss.v043.i05>.}
}

\author{
Klaus Nordhausen, Jyrki Mottonen and Hannu Oja

Maintainer: \packageMaintainer{MNM}

}

\keyword{package }

