% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPredictionsMajority.R
\name{newPredictionsMajority}
\alias{newPredictionsMajority}
\title{Predict tumor subtype for new samples with Majority classifier}
\usage{
newPredictionsMajority(
  createdModelsMajority,
  countDataNew,
  outputDir = paste0("./", format(as.Date(Sys.Date(), "\%Y-\%m-\%d"), "\%Y_\%m_\%d")),
  saveModel = T,
  correctRibo = T
)
}
\arguments{
\item{createdModelsMajority}{R-object containing the rotations and scalings
for each training data subset ($rotationsAndScalingList),
the model to correct for the ribodepletion efficacy ($riboModelList),
which samples were present in each subset ($samplesTrainDefList),
which RNA-transcripts were considered for transformation in the analysis ($nonZeroGenes),
the metadata file associated to the reference cohort ($metaDataRef),
and metadata for the performed run ($metaDataRun).}

\item{countDataNew}{Matrix containing the RNA-transcript per million data for the new samples to be classified.
Samples are in the columns, different RNA-transcripts in the rows.}

\item{outputDir}{Directory in which you would like to store the R-object containing the results. Default is today's date.}

\item{saveModel}{Do you want to save the resulting predictions in an R object? Default is TRUE.}

\item{correctRibo}{Do you want to perform a correction for the ribodepletion protocol on your dataset? Default is TRUE.}
}
\value{
R-object containing the final classifications ($classifications) for the samples,
and the probabilities associated to the different classifications ($probability).
}
\description{
Function to generate predictions for new input samples from their
RNA-seq count data.
}
