% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPredictionsMinority.R
\name{newPredictionsMinority}
\alias{newPredictionsMinority}
\title{Title}
\usage{
newPredictionsMinority(
  createdModelsMinority,
  countDataNew,
  outputDir,
  saveModel = F
)
}
\arguments{
\item{saveModel}{Do you want to save the resulting predictions in an R object?}
}
\description{
Title
}
