% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainPredictionMajorityClassifier.R
\name{obtainPredictionMajorityClassifier}
\alias{obtainPredictionMajorityClassifier}
\title{Obtain predictions from generated majority models}
\usage{
obtainPredictionMajorityClassifier(
  rotationsAndScalingsList,
  dataTrain,
  dataTest,
  metaDataRef,
  samplesTrainDefList,
  testSamples,
  classColumn = classColumn,
  maxNeighbours = 25,
  nModels = 100
)
}
\arguments{
\item{rotationsAndScalingsList}{List containing both the derived PCA-transformation information ($prList)
and the information needed to scale new sample input data to center the features around 0 ($scaleFeaturesList).}

\item{dataTrain}{Data to be used for training, containing the whole reference cohort.}

\item{dataTest}{Data for which the classifier should predict a label.}

\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis.}

\item{samplesTrainDefList}{List of different training data subsets used for the majority voting system.}

\item{testSamples}{IDs for test samples from the test data.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{maxNeighbours}{What is the maximum number of neigbours to be used for the weighted \emph{k}-nearest neighbor algorithm?}

\item{nModels}{How many models should be created for the majority voting system?}
}
\value{
Dataframe containing the predictions for the \emph{nModels} different generated models, with the different folds
in the columns and the different samples to be predicted in the rows.
}
\description{
This function looks at the accompanying prediction for the test samples for the different generated Majority Classifier models.
}
