% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainTrainData.R
\name{obtainTrainData}
\alias{obtainTrainData}
\title{Obtain training data subsets}
\usage{
obtainTrainData(
  metaDataRef,
  classColumn,
  maxSamplesPerType = 50,
  nModels = 100
)
}
\arguments{
\item{metaDataRef}{Metadata file containing the links between the samples and the tumor domain, type and subtype diagnosis.}

\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}

\item{maxSamplesPerType}{How many samples should we maximally use per tumor subtype?}

\item{nModels}{How many models should be created for the classifier?}
}
\value{
A list with the specified number of different training data subsets from the available complete training data,
all with the specified maximum samples per tumor subtype.
}
\description{
This function is used to generate different training subsets for the different models within the M&M setup.
}
