\name{plot.mvl1lm}
\Rdversion{1.1}
\alias{plot.mvl1lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Residual Plot for an mvl1lm Object 
}
\description{
Scatterplots of fitted vs. residual values of the response variable for an mvl1lm object. 
}
\usage{
\method{plot}{mvl1lm}(x, captation = "Residuals vs fitted", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
an object of class \code{mvl1lm}.
}
  \item{captation}{
captation of the figure.
}
  \item{\dots}{
optional plotting arguments.
}
}

\author{Klaus Nordhausen}

\keyword{methods}
