% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot5WithErrorBars.R
\name{plot5WithErrorBars}
\alias{plot5WithErrorBars}
\title{Plot comparison between classifiers on total results}
\usage{
plot5WithErrorBars(
  minorityDir,
  majorityDir,
  minorityDirTest,
  majorityDirTest,
  metaDataTest = NA,
  otherDataSets,
  plotResults = F
)
}
\arguments{
\item{minorityDir}{Directory in which the minority model(s) are stored for the cross-validation setup.}

\item{majorityDir}{Directory in which the majority model(s) are stored for the cross-validation setup.}

\item{minorityDirTest}{Directory in which the minority model(s) are store for the test setup.}

\item{majorityDirTest}{Directory in which the majority model(s) are store for the test setup.}

\item{metaDataTest}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the test set.}

\item{otherDataSets}{List containing a description of the other datasets, with their actual data ($data),
whether we're looking at the tumor type of subtype ($subtype),
what is the probability score threshold that should be used for the M&M classifications ($probabilityScoreThreshold),
whether we're looking at cross-validation or test set results ($TrainOrTest, should be "Train" or "Test"),
and what is the name that you would like to attach to the subset of the data ($subsetName).}

\item{plotResults}{Do you want to obtain the data or (FALSE) or get the resulting plot (TRUE)?}
}
\value{
If plotResults == F, a dataframe containing the accuracy ($meanAccuracy), precision ($meanPrecision),
recall ($meanRecall), and their standard deviations for M&M and other classifirs on a subset of the data ($Subset)
with a specific number of samples ($numberSamples), either within the training data or test set ($TrainOrTest) will be returned.
If plotResults == T, a plot showing the accuracy, precision and recall for M&M and other classifiers for specific subsets of the dataset will be returned.
Included are error bars for the results of the 10 seeds used for the cross-valiation results.
}
\description{
With this function, the overall results will be displayed for the comparison between different classifiers.
Error bars will be generated based on different cross-validation runs.
}
