% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot5withErrorBarsSubsampling.R
\name{plot5withErrorBarsSubsampling}
\alias{plot5withErrorBarsSubsampling}
\title{Plot comparison between classifiers on subsampled results}
\usage{
plot5withErrorBarsSubsampling(
  minorityDir,
  majorityDir,
  minorityDirTest,
  majorityDirTest,
  metaDataTest = NA,
  otherDataSets,
  subsamplePercentage,
  numberSubsamples = 100,
  returnPlot = F,
  probabilityScoreMinimum = 0
)
}
\arguments{
\item{minorityDir}{Directory in which the minority model(s) are stored for the cross-validation setup.}

\item{majorityDir}{Directory in which the majority model(s) are stored for the cross-validation setup.}

\item{minorityDirTest}{Directory in which the minority model(s) are store for the test setup.}

\item{majorityDirTest}{Directory in which the majority model(s) are store for the test setup.}

\item{metaDataTest}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the test set.}

\item{otherDataSets}{List containing a description of the other datasets, with their actual data ($data),
whether we're looking at the tumor type of subtype ($subtype),
what is the probability score threshold that should be used for the M&M classifications ($probabilityScoreThreshold),
whether we're looking at cross-validation or test set results ($TrainOrTest, should be "Train" or "Test"),
and what is the name that you would like to attach to the subset of the data ($subsetName).}

\item{subsamplePercentage}{Which percentage of samples would you want to subsample each time?}

\item{numberSubsamples}{How often do you want to subsample?}

\item{returnPlot}{Do you want to obtain the data or (FALSE) or get the resulting plot (TRUE)?}

\item{probabilityScoreMinimum}{Do you want to say that classifications with a very low probability score should be called 'wrong classifications'?
If yes, please specify the probability score this should happen at. This will be used within the results for the methylation classifiers.}

\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}
}
\value{
If returnPlot == F, a dataframe containing the accuracy ($meanAccuracy), precision ($meanPrecision),
recall ($meanRecall), and their standard deviations for M&M and other classifirs on a subset of the data ($Subset)
with a specific number of samples ($numberSamples), either within the training data or test set ($TrainOrTest) will be returned.
If returnPlot == T, a plot showing the accuracy, precision and recall for M&M and other classifiers for specific subsets of the dataset will be returned.
Included are error bars for the results of the 10 seeds used for the cross-valiation results.
}
\description{
With this function, the results of will be displayed for the comparison between different classifiers.
Error bars will be generated based on results for 100 different data subsamples.
}
