% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAccuraciesPerTumorTypeSize.R
\name{plotAccuraciesPerTumorTypeSize}
\alias{plotAccuraciesPerTumorTypeSize}
\title{Plot the accuracies of the filtered different tumor type predictions per size block}
\usage{
plotAccuraciesPerTumorTypeSize(fractionsCorrect)
}
\arguments{
\item{fractionsCorrect}{Dataframe containing the results per tumor type size block obtained from function 'getAccuraciesPerTumorTypeSize'.}
}
\value{
ggplot object that contains the barplot with the accuracies for the
tumor types of different sample bins (e.g. only 3 available samples per tumor type is a bin).
}
\description{
Plot the accuracies of the filtered different tumor type predictions per size block
}
