% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAccuraciesPerTumorTypeSizeNoFilter.R
\name{plotAccuraciesPerTumorTypeSizeNoFilter}
\alias{plotAccuraciesPerTumorTypeSizeNoFilter}
\title{Plot the accuracies of the different tumor type predictions per size block}
\usage{
plotAccuraciesPerTumorTypeSizeNoFilter(fractionsCorrect, MnM = F)
}
\arguments{
\item{fractionsCorrect}{Dataframe containing the results per tumor type size block.}
}
\value{
ggplot object that contains the barplot with the accuracies for the
tumor types of different sample bins (e.g. only 3 available samples per tumor type is a bin).
No filtering was applied.
}
\description{
Plot the accuracies of the different tumor type predictions per size block
}
