% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortDomain.R
\name{plotCohortDomain}
\alias{plotCohortDomain}
\title{Plot UMAP Domains}
\usage{
plotCohortDomain(dataUMAPList, useLabels = T)
}
\arguments{
\item{useLabels}{Do you want to use labels within your plot (= TRUE) or only show
the different domains with different colors (= FALSE)? Default = TRUE.}

\item{dataUMAP}{Dataframe containing the UMAP coordinates, the tumor type ($subclass),
the domain ($Domain) and the desired abbreviation for the tumor type ($abbreviation)}
}
\value{
ggplot with all datapoints from the reference cohort,
color coded by the domain.
}
\description{
Function to plot the UMAP coordinates for the different domains.
All domains are colored differently to
distinguish them from one another.
}
