% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortTumorType.R
\name{plotCohortTumorType}
\alias{plotCohortTumorType}
\title{Plot UMAP tumor types}
\usage{
plotCohortTumorType(
  dataUMAP,
  domain,
  tumorType,
  plotColors,
  classColumn,
  subtype = F,
  abbreviations,
  useLabels = T,
  useManualColors = F
)
}
\arguments{
\item{dataUMAP}{Dataframe containing the UMAP coordinates, the tumor type ($subclass),
the domain ($Domain) and the desired abbreviation for the tumor type ($abbreviation)}

\item{domain}{Which domain do you want to plot the tumor types for?}

\item{plotColors}{Which colors do you want to use for the tumor types?}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{subtype}{Do you want to visualize the subtypes of one tumor type?
#' @param tumorType If subtype == T, which tumor type would you want to visualize?}

\item{abbreviations}{Abbreviations to be used in the plot for the tumor (sub)types.}

\item{useLabels}{Do you want to supply labels within the plot?}

\item{useManualColors}{Do you want to supply the colors to be used within the UMAP for the labels?}
}
\value{
ggplot with only the datapoints of the selected domain,
color coded by the tumor type.
}
\description{
Function to plot the UMAP coordinates for the different tumor type samples
belonging to one domain. All tumor types are colored differently to
distinguish them from one another.
}
