% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortUMAP.R
\name{plotCohortUMAP}
\alias{plotCohortUMAP}
\title{Plot UMAP coordinates colored by tumor domain, tumor types or tumor subtypes}
\usage{
plotCohortUMAP(
  dataUMAPList,
  domain = NA,
  subtype = F,
  tumorType = NA,
  plotColors = NA,
  useLabels = T
)
}
\arguments{
\item{dataUMAPList}{List containing the UMAP coordinate data
and the desired abbreviation for the tumor type ($abbreviation), resulting from the function 'createUMAPcohort'}

\item{domain}{Which domain do you want to plot the tumor types for? If not specified, visualization will be performed on the domain level by default.}

\item{subtype}{In case domain is specified: Do you want to visualize on the tumor subtype level (TRUE) or on the tumor type level (FALSE)?}

\item{tumorType}{If domain is specified and subtype = TRUE, there's a possibility to plot the tumor subtypes for only one tumor type.
Please specify which tumor type you would like to visualize within this parameter.}

\item{plotColors}{Which colors do you want to use for the tumor types? If not specified, default ggplot2 colors will be used.}

\item{useLabels}{Do you want to supply labels within the plot?
If that would be desired, fill out the tumor type that you would like to visualize at 'tumorType' (for example: "B-ALL").}
}
\value{
ggplot with only the datapoints of the selected domain,
color coded by the tumor type (subtype = F) or tumor subtype (subtype = T).
}
\description{
Function to plot the UMAP coordinates for the different tumor (sub)type samples
belonging to one domain or tumor type.
}
