% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComparisonBetweenClassifiers.R
\name{plotComparisonBetweenClassifiers}
\alias{plotComparisonBetweenClassifiers}
\title{Plot comparison between classifiers}
\usage{
plotComparisonBetweenClassifiers(comparisonDF, subsampling)
}
\arguments{
\item{comparisonDF}{Dataframe containing the accuracy ($meanAccuracy), precision ($meanPrecision),
recall ($meanRecall), and their standard deviations for M&M and other classifirs on a subset of the data ($Subset)
with a specific number of samples ($numberSamples), either within the training data or test set ($TrainOrTest).}

\item{subsampling}{Did you use subsampling to generate your comparisonDF?}
}
\value{
plot showing the accuracy, precision and recall for M&M and other classifiers for specific subsets of the dataset will be returned.
Included are error bars for the results based on either different seeds used for the cross-valiation results or subsampling.
}
\description{
Plot comparison between classifiers
}
