% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfusionMatrix.R
\name{plotConfusionMatrix}
\alias{plotConfusionMatrix}
\title{Plot tumor type or subtype confusion matrix}
\usage{
plotConfusionMatrix(
  domain = NA,
  confusionPlotInfo,
  domainColor = NA,
  colorTiles = "#012695"
)
}
\arguments{
\item{domain}{Which domain do you want to plot? This name should be available in the
domain column of the abbreviations. If not specified, plot will be made for all tumor domains.}

\item{confusionPlotInfo}{List resulting from running the function 'getConfusionMatrixPlot'.}

\item{domainColor}{If one domain is plotted, which colors should we use for each designated tumor type?
If not specified, default ggplot colors will be used.}

\item{colorTiles}{Which domain color tiles do we want to use? Only applicable when argument \emph{domain} is specified.}
}
\value{
ggplot object containing the tumor type or subtype CM plot.
}
\description{
Function to plot the confusion matrix (CM) plot within all domains (domain = NA)
or within a single domain (e.g. domain = "Hemato").
The CM plots can both be made on the tumor type and subtype level,
which depends on the information stored within the 'confusionPlotInfo'.
The CM plot shows the reference-classification combinations for the tumor (sub)types.
Within tumor subtype CM plots, it is shown which tumor subtypes belong to the same parent tumor type using color coding.
}
