% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfusionMatrixPerDomain.R
\name{plotConfusionMatrixPerDomain}
\alias{plotConfusionMatrixPerDomain}
\title{Plot tumor subtype confusion matrix}
\usage{
plotConfusionMatrixPerDomain(
  domain,
  confusionPlotInfo,
  domainColor,
  colorTiles = "#012695"
)
}
\arguments{
\item{domain}{Which domain do you want to plot? This name should be present in the
domainColorumn of the metadata.}

\item{domainColor}{Which colors should we use for each designated tumor type?}

\item{colorTiles}{Which domain color do we want to use?}
}
\value{
ggplot object containing the tumor subtype confusion matrix.
}
\description{
Function to plot the confusion matrix plot for a single domain,
showing the reference-prediction combinations for the tumor subtypes.
Using a color coding, it is shown which tumor subtypes belong to the same overheading tumor type.
IMPORTANT: The order of the colors will be applied to the tumor types in the order of the
tumors within the abbreviations dataframe. Make sure you specify the
order of the tumor types well within the abbreviations, not only concerning the domains,
but also the subsequent tumor types order and tumor subtypes order.
}
