% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotF1.R
\name{plotF1}
\alias{plotF1}
\title{Plot F1-scores for tumors from different population frequencies}
\usage{
plotF1(dataMeanF1, separateMeanF1, subtype)
}
\arguments{
\item{dataMeanF1}{Dataframe obtained from function 'calculateMeanAndSDAccuracy'.}

\item{separateMeanF1}{Dataframe obtained from function 'calculateSeparateF1' .}

\item{subtype}{Do you want to visualize the tumor subtype (TRUE) or tumor type (FALSE)?
dataMeanF1 and separateMeanF1 should have been generated on the same level,
being either subtype or tumor type.}
}
\value{
Plot showing the average F1 values for tumors stratified by population frequency,
and separate F1 values for the separate tumors within the population frequency blocks.
}
\description{
Plot F1-scores for tumors from different population frequencies
}
