% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerformanceDomain.R
\name{plotPerformanceDomain}
\alias{plotPerformanceDomain}
\title{Function to obtain (plot for) perfomance within the different domains}
\usage{
plotPerformanceDomain(
  precisionDomainTrain,
  precisionDomainTest,
  levelsDomain,
  returnPlot
)
}
\arguments{
\item{precisionDomainTrain}{Dataframe containing results for reference cohort data obtained from the function 'precisionsForDomains'}

\item{precisionDomainTest}{Dataframe containing results for test set data obtained from the function 'precisionsForDomains'}

\item{levelsDomain}{What are the names you have within the domains?}

\item{returnPlot}{Do you want to obtain the data or (FALSE) or get the resulting plot (TRUE)?}
}
\value{
If returnPlot = T: Plot with the performance within the different domains.
If returnPlot = F: Dataframe containing the mean accuracy, precision and recall within the different domains.
}
\description{
Function to obtain (plot for) perfomance within the different domains
}
