% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPieChartDomain.R
\name{plotPieChartDomain}
\alias{plotPieChartDomain}
\title{Plot pie chart from one domain}
\usage{
plotPieChartDomain(
  data,
  domain,
  classColumn,
  higherClassColumn,
  domainColumn,
  textSizeClass = 0.5,
  textSizeSubspec = 0.45,
  freqSameTumorType,
  plotColors,
  saveImage = T,
  storeLocation
)
}
\arguments{
\item{data}{Dataframe containing the fractions, angles and positions of the different labels,
with the labels being present on both the classColumn and higherClassColumn level.}

\item{domain}{Which domain do you want to plot? This name should be present in the
Domain column of the domainColumn of the metadata.}

\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}

\item{domainColumn}{Column in the metadata file that contains the domain labels.}

\item{textSizeClass}{What size should the letters have for the tumor type?}

\item{textSizeSubspec}{What size should the letters have for the tumor subtype?}

\item{freqSameTumorType}{How many subtypes belong to one tumor type? In order of the tumor types.}

\item{plotColors}{Which colors do you want to use to color your plot?}

\item{saveImage}{Do you want to save your image? Boolean (T/F) input.}

\item{storeLocation}{Where do you want to save the image?}
}
\value{
Piechart of how many samples are present within the dataset belonging to a
certain tumor type, and to a certain tumor subtype.
The piechart is generated for one domain, as there are many different tumor types and subtypes
within each domain.
}
\description{
This function is dedicated to plotting the pie chart from one domain.
}
