% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPRPlotGlobal.R
\name{plotPRPlotGlobal}
\alias{plotPRPlotGlobal}
\title{Plot precision-recall curves of classifier comparison - global}
\usage{
plotPRPlotGlobal(dataPR, otherClassifierName)
}
\arguments{
\item{dataPR}{Dataframe containing the precision and recall values at specific cutoffs,
for both the train and test set. Also, a minimum and maximum precision are supplied
to generate a shadow of the fluctuations.}

\item{otherClassifierName}{What is the name of the other classifier within dataPR?}

\item{removeLegend}{Do you want to remove the legend?}
}
\value{
Precision-recall curve for train and test dataset of M&M and other classifier, with a shadow for cross-validation.
This function shows the full range from 0-1 for the precision values.
}
\description{
Plot precision-recall curves of classifier comparison - global
}
