% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPRPlotZoomed.R
\name{plotPRPlotZoomed}
\alias{plotPRPlotZoomed}
\title{Plot precision-recall curves of classifier comparison - zoom}
\usage{
plotPRPlotZoomed(dataPR, otherClassifierName, removeLegend = T)
}
\arguments{
\item{dataPR}{Dataframe containing the precision and recall values at specific cutoffs,
for both the train and test set. Also, a minimum and maximum precision are supplied
to generate a shadow of the fluctuations.}

\item{otherClassifierName}{What is the name of the other classifier within dataPR?}

\item{removeLegend}{Do you want to remove the legend?}
}
\value{
Precision-recall curve for train and test dataset of M&M and other classifier, with a shadow for cross-validation.
This function shows a close-up of the high precision range.
}
\description{
Plot precision-recall curves of classifier comparison - zoom
}
