% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleHeterogeneity.R
\name{plotSampleHeterogeneity}
\alias{plotSampleHeterogeneity}
\title{Title}
\usage{
plotSampleHeterogeneity(
  tumorHeterogeneityTrain,
  tumorHeterogeneityTest,
  subtype
)
}
\arguments{
\item{subtype}{}
}
\description{
Title
}
