% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeparateClassifierAccuracies.R
\name{plotSeparateClassifierAccuracies}
\alias{plotSeparateClassifierAccuracies}
\title{Title}
\usage{
plotSeparateClassifierAccuracies(fractionCorrectTotalPivoted)
}
\arguments{
\item{fractionCorrectTotalPivoted}{}
}
\description{
Title
}
