% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTop3Accuracy.R
\name{plotTop3Accuracy}
\alias{plotTop3Accuracy}
\title{Plot accuracy within the top 3 highest scoring classification labels M&M}
\usage{
plotTop3Accuracy(meanAndSDPlotTrain, meanAndSDPlotTest, metaDataRef, subtype)
}
\arguments{
\item{meanAndSDPlotTrain}{Dataframe containing the average performance of tumor classifications within a certain frequency range (nCases) for the reference cohort.
Object obtained from running the function calculateMeanAndSDAccuracy}

\item{meanAndSDPlotTest}{Dataframe containing the average performance of tumor classifications within a certain frequency range (nCases) for the test set.
Object obtained from running the function calculateMeanAndSDAccuracy}

\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the reference cohort.}

\item{subtype}{Do you want to analyse the results on the tumor subtype classification level?}
}
\value{
Plot showing the accuracy if not only the top scoring classification label, but also the second and third highest labels are taken into consideration.
Plot is segregated into the different class frequencies.
}
\description{
Plot accuracy within the top 3 highest scoring classification labels M&M
}
