% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precisionRecallPlotData.R
\name{precisionRecallPlotData}
\alias{precisionRecallPlotData}
\title{Generate data for PR-plot comparison M&M v.s. others}
\usage{
precisionRecallPlotData(
  minorityDir,
  majorityDir,
  predictionsMMTest,
  otherClassifierResultsTrain,
  otherClassifierResultsTest,
  otherClassifierName,
  originalCallColumnOtherClassifier,
  otherClassifierPredictionColumn,
  scoreName,
  nSeeds = 10,
  subtype,
  metaDataRef,
  nModels,
  classColumn,
  higherClassColumn
)
}
\arguments{
\item{minorityDir}{Directory in which the minority model(s) are stored for the cross-validation setup.}

\item{majorityDir}{Directory in which the majority model(s) are stored for the cross-validation setup.}

\item{predictionsMMTest}{Classifications of M&M algorithm on the test set, obtained from the function integrateMM.}

\item{otherClassifierResultsTrain}{Results for other classifier within the reference cohort.}

\item{otherClassifierResultsTest}{Results for other classifier within the test set.}

\item{otherClassifierName}{What is the name of the other classifier?}

\item{originalCallColumnOtherClassifier}{What column in the results for the other classifier contains the original diagnoses for the samples?}

\item{otherClassifierPredictionColumn}{What column in the results for the other classifier contains the classifications?}

\item{scoreName}{What column in the results for the other classifier contains the probability scores?}

\item{nSeeds}{How many seeds was the cross-validation setup run with?}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?}

\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the reference cohort.}

\item{nModels}{How many models should be created for the majority voting system?}

\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}
}
\value{
Dataframe containing the necessary data points to generate PR-plots for M&M versus other classifiers.
}
\description{
Generate data for PR-plot comparison M&M v.s. others
}
