% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precisionsForDomains.R
\name{precisionsForDomains}
\alias{precisionsForDomains}
\title{Compare results between samples from different domains}
\usage{
precisionsForDomains(
  minorityDir,
  majorityDir,
  subtype,
  metaDataTest,
  probabilityThreshold,
  metaDataRef
)
}
\arguments{
\item{minorityDir}{Directory in which the minority model(s) are stored.}

\item{majorityDir}{Directory in which the majority model(s) are stored.}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?}

\item{metaDataTest}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the test set.
This file needs to have the same column names for the tumor subtype labels (classColumn),
tumor type labels (higherClassColumn) and domain labels (domainColumn).}

\item{probabilityThreshold}{What is the probability score threshold you would like to use to call a classification 'confident' for M&M?}

\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis.}
}
\value{
Dataframe containing the mean accuracy ($meanAccuracy),
precision ($meanPrecision), and recall ($meanRecall),
together with their standard deviations ($sdAccuracy,
$sdPrecision, $sdRecall), and the total number of samples associated to the domain group ($nSamples)
}
\description{
Compare results between samples from different domains
}
