% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRiboCounts.R
\name{predictRiboCounts}
\alias{predictRiboCounts}
\title{Correct new count data for ribodepletion efficacy}
\usage{
predictRiboCounts(riboModel, data)
}
\arguments{
\item{riboModel}{Model for ribocorrection, normally stored inside the R-object
obtained from running the function createModelsMinority or createScalingsMajority as 'riboModelList'.}

\item{data}{Count data to be transformed.}
}
\value{
Count data corrected for the efficacy of the ribodepletion protocol.
}
\description{
This function is designed to use the same ribodepletion correcting model
on the new samples as was used on the reference cohort.
}
