% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryRecurrenceMetastasisNumbers.R
\name{primaryRecurrenceMetastasisNumbers}
\alias{primaryRecurrenceMetastasisNumbers}
\title{Get performance measures for separate heterogeneous groups within data set}
\usage{
primaryRecurrenceMetastasisNumbers(
  metaDataRef,
  predictionsMM,
  probabilityScore,
  columnOfInterest,
  labelTypes,
  includeAllRow = F
)
}
\arguments{
\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the reference cohort.}

\item{predictionsMM}{Dataframe containing the top 3 final tumor type classification labels ($predict{2,3}) with their accompanying probability scores
($probability{1,2,3} and the original diagnosis label ($originalCall).}

\item{probabilityScore}{What is the probability score threshold you would like to use to call a classification 'confident'?}

\item{columnOfInterest}{Which column contains the different groups that should be analyzed?}

\item{labelTypes}{What are the names for the different groups within the column of interest (e.g. Primary, Recurrence, Metastasis)?}

\item{includeAllRow}{Do you want to also generate an entry that specifies the performance for all samples taken together?}
}
\value{
Dataframe containing the total number of samples within the group ($numberSamples), how many of those were correct ($correctSamples),
how many errors were present ($errorsSamples),
what was the accuracy, how many samples were confidently classified ($numberSamplesFiltered),
how many of those confident classifications were correct ($errorsFiltered),
what was the accuracy within the confident classifications ($precision) and how many samples reached the confidence treshold ($recall).
}
\description{
Get performance measures for separate heterogeneous groups within data set
}
