\name{print.anovamvl1lm}
\Rdversion{1.1}
\alias{print.anovamvl1lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Printing an Object of Class anovamvl1lm
}
\description{
Printing an object of class 'anovamvl1lm'.
}
\usage{
\method{print}{anovamvl1lm}(x, ...)
}

\arguments{
  \item{x}{
an object of class \code{anovamvl1lm}.
}
  \item{\dots}{
needed for other methods.
}
}
\author{Klaus Nordhausen, \email{klaus.nordhausen@uta.fi} }

\keyword{methods}
