\name{print.mvcloc}
\alias{print.mvcloc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Printing an 'mvcloc' Object }
\description{
 Printing an mvcloc object.
}
\usage{
\method{print}{mvcloc}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{an object of class \code{mvcloc}.}
  \item{\dots}{arguments that can be passed further on.}
}


\author{Jyrki Mottonen \email{jyrki.mottonen@helsinki.fi}}

\keyword{print}
