\name{print.mvl1lm}
\Rdversion{1.1}
\alias{print.mvl1lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Printing an mvl1lm Object 
}
\description{
Printing of an mvl1lm object.
}
\usage{
\method{print}{mvl1lm}(x, digits = 3, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
an object of class \code{mvl1lm}.
}
  \item{digits}{
minimal number of _significant_ digits.
}
  \item{\dots}{
needed for other methods.
}
}

\author{Klaus Nordhausen}

\keyword{methods}

