\name{print.mvloc}
\alias{print.mvloc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Printing an 'mvloc' Object }
\description{
 Printing an mvloc object.
}
\usage{
\method{print}{mvloc}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{an object of class \code{mvloc}.}
  \item{\dots}{arguments that can be passed further on.}
}


\author{Klaus Nordhausen}

\keyword{print}
