\name{print.mvPCA}
\alias{print.mvPCA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Printing Method for a Principal Component Object of Type mvPCA}
\description{Prints an object of class mvPCA.}
\usage{
\method{print}{mvPCA}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{object of type 'mvPCA' }
  \item{\dots}{ needed for other printing methods.}
}

\author{Klaus Nordhausen}

\keyword{print}
\keyword{methods}
