% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCountData.R
\name{processCountData}
\alias{processCountData}
\title{Function used to process RNA-Seq count data from a text file and save as a CSV
This data is obtained from a Diagnostics RNA-Seq based Pipeline}
\usage{
processCountData(input_file_path, output_file_path, sample_name)
}
\arguments{
\item{input_file_path}{String representing the path to the input RNA-Seq count data file.}

\item{output_file_path}{String representing the path where the output CSV file will be saved.}

\item{sample_name}{String representing the name of the column for which CPM values are calculated, this is the sample_name}
}
\value{
None. The function saves the processed data as a CSV file to the specified path.
}
\description{
This function processes RNA-Seq count data from a specified input file, summarizes the
counts by gene, and outputs the summarized data as a CSV file. The data is grouped by
gene name, and CPM values are calculated as the sum of counts per gene.
}
