% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRPlotComparison.R
\name{PRPlotComparison}
\alias{PRPlotComparison}
\title{Title}
\usage{
PRPlotComparison(
  predictionsMM,
  otherClassifierName,
  otherClassifierResults,
  originalCallColumnOtherClassifier,
  otherClassifierPredictionColumn,
  scoreName
)
}
\arguments{
\item{predictionsMM}{Predictions for samples by M&M, coming from the function integrateMM.}

\item{otherClassifierName}{What is the name of the other classifier?}

\item{otherClassifierResults}{Dataframe containing the clsasifications by the other classifier, together with their probability score.}

\item{originalCallColumnOtherClassifier}{Column within otherClassifierResults containing the original diagnosis label}

\item{otherClassifierPredictionColumn}{Column within otherClassifierReults containing the prediction}

\item{scoreName}{Column within otherClassifierResults containing the probability score.}
}
\description{
Title
}
