% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceFeatures.R
\name{reduceFeatures}
\alias{reduceFeatures}
\title{Reduce features Minority classifier with Random Forest importance score}
\usage{
reduceFeatures(
  dataTrain,
  samplesTrainDefList,
  ntree = 500,
  nModels = 10,
  nFeatures = 300,
  nANOVAgenes
)
}
\arguments{
\item{dataTrain}{Data to be used for model training, containing the all samples available within the training dataset.}

\item{samplesTrainDefList}{List of different training data sample subsets used for subsetting the available training dataset (dataTrain).}

\item{ntree}{How many trees should we generate during the weighted RF procedure?}

\item{nModels}{How many models should be created for the majority voting system?}

\item{nFeatures}{How many features should we keep after determining
the most important RNA-transcripts using the RF Importance Score?}

\item{nANOVAgenes}{How many RNA-transcripts did we select during the ANOVA procedure?}
}
\value{
Names of the top most important RNA-transcripts present in the training data,
containing most information to distinguish the different tumor (sub)types.
}
\description{
This functions' purpose is to further reduce the number of features within the Minority classifier.
Ten Random Forest (RF) models will be generated using the reference cohort training data,
from which the importance scores for the different RNA-transcripts will be extracted.
Each model is generated using a different subset of the training data, as specified in the \emph{samplesTrainDefList}.
The \emph{nFeatures} RNA-transcripts with the highest importance scores will be selected for the final Minority Classifier training.
}
