% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBiasTestSet.R
\name{removeBiasTestSet}
\alias{removeBiasTestSet}
\title{Remove bias from test set}
\usage{
removeBiasTestSet(metaDataRef, metaDataTest, classColumn, patientIDColumn)
}
\arguments{
\item{classColumn}{Which column in the metadata file should be used to filter out data points
from the test set that do not have a corresponding label?}

\item{newRefCohort}{Reference cohort R-object containing both the RNA-seq count data (refCohort$counts) and
the metadata associated to this count data (refCohort$metaData).}

\item{testSet}{Testset R-object containing both the RNA-seq count data (testSet$counts) and
the metadata associated to this count data (testSet$metaData).}
}
\value{
The test set without samples belonging to patients from the reference cohort.
}
\description{
Remove samples belonging to patients that are present in the reference cohort from the test set,
and are thereby violating the assumption of independence between the reference cohort and test set.
Also, the samples that can never be predicted correctly (labels not within reference cohort) are removed.
}
