% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDuplicates.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove duplicate entries from patients}
\usage{
removeDuplicates(refCohort, inputDir)
}
\arguments{
\item{refCohort}{Reference cohort R-object containing both the RNA-seq count data (refCohort$counts) and
the metadata associated to this count data (refCohort$metaData).}

\item{inputDir}{Directory where the sample information is stored. This information is needed to determine if
multiple samples from the same patient within the same diagnosis are used.}
}
\value{
The reference cohort cleaned from duplicate samples, with one sample per patient per diagnosis.
This means that if one patient has multiple tumors, multiple samples with the same patient ID can still be present within the dataset.
}
\description{
Function that ensures that there is only one sample per patient per tumor type present within the reference cohort,
to prevent selection bias.
}
