\name{residuals.mvl1lm}
\Rdversion{1.1}
\alias{residuals.mvl1lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Residuals of an mvl1lm Object
}
\description{
Extracts the residuals of an mvl1lm object.
}
\usage{
\method{residuals}{mvl1lm}(object, ...)
}
\arguments{
  \item{object}{
an object of class \code{mvl1lm}.
}
  \item{\dots}{
needed for other methods.
}
}

\author{Klaus Nordhausen, \email{klaus.nordhausen@uta.fi} }

\keyword{methods}
