% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboCorrectCounts.R
\name{riboCorrectCounts}
\alias{riboCorrectCounts}
\title{Correct ribodepletion protocol}
\usage{
riboCorrectCounts(data, proteinCodingGenes, outputDir, saveRiboModels = T)
}
\arguments{
\item{data}{Dataframe containing the ribo-depleted count data from an RNA-seq experiment.}

\item{proteinCodingGenes}{Names of the protein-coding genes within our RNA-seq count data.}

\item{outputDir}{Directory where the generated model for predicting the protein fraction should be stored.}

\item{saveRiboModels}{Do you want to save the riboModel?}
}
\value{
list containing the corrected count data ($counts) and the essentials needed to correct new data using the same model ($modelList).
}
\description{
This functions is used to account for the differences in efficacy of the ribodepletion protocol.
It is estimated what is the percentage of protein coding transcripts within the total transcriptome, and subsequently
the protein coding fraction is scaled up to 100\%, simultaneously removing certain residual ribosomal RNAs from the dataset.
}
