% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAnovaGenes.R
\name{selectAnovaGenes}
\alias{selectAnovaGenes}
\title{Select most informative genes Minority Classifier}
\usage{
selectAnovaGenes(metaDataRef, countDataRef, nANOVAgenes, classColumn)
}
\arguments{
\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis for the reference cohort.}

\item{countDataRef}{Matrix containing the RNA-transcript per million data. Patients are in the columns,
different genes in the rows.}

\item{nANOVAgenes}{How many genes should we select using the F-statistic of ANOVA?}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}
}
\value{
Vector containing the names of the most interesting genes according to the F-statistic of ANOVA.
}
\description{
This function allows for the extraction of the genes that are most differently expressed between the different tumor (sub)types.
The gene selection is based on an analysis of variance (ANOVA) with equal variances, where the genes with the highest F-scores are selected.
}
