% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectData.R
\name{selectData}
\alias{selectData}
\title{Selecting tumor (sub)types}
\usage{
selectData(refCohort, classColumn, n = 3)
}
\arguments{
\item{refCohort}{Reference cohort R-object containing both the RNA-seq count data (refCohort$counts) and
the metadata associated to this count data (refCohort$metaData).}

\item{classColumn}{Column in the metadata where the prediction labels for the classification process are specified.}

\item{n}{Specification of the minimum number of samples that should be present for each tumor (sub)type.}
}
\value{
The new reference cohort containing only the tumor (sub)types with more than \emph{n} entries within the dataset.
}
\description{
This function removes all tumor (sub)types from the dataset which do not have \emph{n} or more samples. The M&M setup
requires an absolute minimum of 3 samples per tumor type for the cross-validation setup,
and 2 for the setup for new samples due to the use of the F-statistic for gene selection.
However, larger minimum numbers can be specified within this function as well.
}
