% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeaturesMajority.R
\name{selectFeaturesMajority}
\alias{selectFeaturesMajority}
\title{Select the most variable features for each Majority Classifier model}
\usage{
selectFeaturesMajority(dataTrain, samplesTrainDefList, nFeatures = 2500)
}
\arguments{
\item{dataTrain}{Data to be used for model training, containing all the samples available within the training dataset.}

\item{samplesTrainDefList}{List of different training data sample subsets used for subsetting the available
training dataset (dataTrain).}

\item{nFeatures}{How many of the most variable RNA-transcripts within the dataset should we select for the
Majority Classifier? (default = 2500)}
}
\value{
List with the same length as the number of models, comprising dataframes with the selected samples
per model
}
\description{
Select the most variable features for each Majority Classifier model
}
