% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartRound.R
\name{smartRound}
\alias{smartRound}
\title{Round all doubles within a vector}
\usage{
smartRound(x, digits = 0)
}
\arguments{
\item{x}{vector with numbers that need to be rounded.}

\item{digits}{number of decimals the vector needs to be rounded to.}
}
\value{
vector with rounded doubles with \emph{digits} decimals.
}
\description{
Function to make sure the combined counts of the normalized RNA-seq data still
add up to a million after correction for the ribodepletion protocol.
}
