\name{summary.mvcloc}
\alias{summary.mvcloc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Summarizing an 'mvcloc' Object }
\description{
 Summarizing an \code{'mvcloc'} object.
}
\usage{
\method{summary}{mvcloc}(object,..., digits = 4)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{an object of class \code{mvcloc}.}
  \item{\dots}{needed for other summary methods.}
  \item{digits}{number of digits for rounding.}
}
\author{Jyrki Mottonen \email{jyrki.mottonen@helsinki.fi}}
\keyword{print}
