\name{summary.mvl1lm}
\Rdversion{1.1}
\alias{summary.mvl1lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary for an mvl1lm Object 
}
\description{
Gives a detailed output for an object of class mvl1lm. Note that the output will differ for different score functions used.
}
\usage{
\method{summary}{mvl1lm}(object, ..., digits = 3)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
an object of class \code{mvl1lm}.
}
  \item{\dots}{
needed for other methods.
}
  \item{digits}{
minimal number of _significant_ digits.
}
}

\author{Klaus Nordhausen, \email{klaus.nordhausen@uta.fi} }

\keyword{methods}

