\name{summary.mvloc}
\alias{summary.mvloc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Summarizing an 'mvloc' Object }
\description{
 Summarizing an \code{'mvloc'} object.
}
\usage{
\method{summary}{mvloc}(object,..., digits = 4)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{an object of class \code{mvloc}.}
  \item{\dots}{needed for other summary methods.}
  \item{digits}{number of digits for rounding.}
}

\author{Klaus Nordhausen, \email{klaus.nordhausen@uta.fi} }

\keyword{print}
