% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFoldCrossValidationMajority.R
\name{tenFoldCrossValidationMajority}
\alias{tenFoldCrossValidationMajority}
\title{10x cross-validation majority classifier}
\usage{
tenFoldCrossValidationMajority(
  countDataRef,
  metaDataRef,
  classColumn,
  patientColumn,
  nModels = 100,
  maxSamplesPerType = 50,
  nFeatures = 2500,
  nComps = 100,
  maxNeighbours = 25,
  whichSeed = 1,
  outputDir = ".",
  proteinDir = "./input/"
)
}
\arguments{
\item{countDataRef}{Matrix containing the RNA-transcript per million data. Patients are in the columns,
different genes in the rows.}

\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{patientColumn}{Column in the metadata file that contains the patient labels.}

\item{nModels}{How many models should be created for the majority voting system?}

\item{maxSamplesPerType}{How many samples should we maximally use per tumor (sub)type?}

\item{nFeatures}{How many of the most variable genes within the dataset should we select for principal component analysis (PCA)?}

\item{nComps}{How many principal components will be selected after PCA?}

\item{maxNeighbours}{What is the maximum number of neigbours to be used for the weighted \emph{k}-nearest neighbor algorithm?}

\item{whichSeed}{For reproducibility, the seed can be specified with this parameter.}

\item{outputDir}{Directory in which you would like to store the R-object containing the results.}

\item{proteinDir}{In which directory can we find the file specifying the names of protein-coding genes within our dataset?}
}
\value{
R-object containing the predictions ($classifications), classifications errors ($wrongClassifications),
the probabilities for each classification ($probabilityList), the metadata file associated to the reference cohort ($metaData),
and metadata for the performed run ($metaDataRun).
}
\description{
Setup to automatically run the 10x stratified cross-validation majority classifier.
}
