% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainKNNClassifier.R
\name{trainKNNClassifier}
\alias{trainKNNClassifier}
\title{Calculate optimal value for k in a given model}
\usage{
trainKNNClassifier(
  rotationsAndScalingsList,
  metaDataRef,
  classColumn,
  maxNeighbors = 25,
  nModels = 100
)
}
\arguments{
\item{rotationsAndScalingsList}{List containing both the derived PCA-transformation information ($prList)
and the information needed to scale new sample input data to center the features around 0 ($scaleFeaturesList).}

\item{metaDataRef}{Metadata file containing the links between the samples and the tumor domain, type and subtype diagnosis.}

\item{classColumn}{Name of column in the metadata file that contains the tumor subtype labels.}

\item{maxNeighbors}{What is the maximum number of neighbors to be used for the weighted \emph{k}-nearest neighbor algorithm?}

\item{nModels}{How many models should be created for the Majority classifier?}
}
\value{
Vector with optimized values for k within the different weighted kNN models
}
\description{
Calculate optimal value for k in a given model
}
