% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatedVSUntreated.R
\name{treatedVSUntreated}
\alias{treatedVSUntreated}
\title{Title}
\usage{
treatedVSUntreated(metaDataRef, predictionsMM, probabilityScore, testSet)
}
\arguments{
\item{probabilityScore}{}
}
\description{
Title
}
