% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumorTypeSubtypePie.R
\name{tumorTypeSubtypePie}
\alias{tumorTypeSubtypePie}
\title{Plot pie-chart area with label and color}
\usage{
tumorTypeSubtypePie(
  x,
  labels = names(x),
  edges = 200,
  radius = 0.8,
  clockwise = FALSE,
  init.angle = if (clockwise) 90 else 0,
  density = NULL,
  angle = 45,
  col = NULL,
  border = NULL,
  lty = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{Data containing the numbers of a specific tumor (sub)type and their names.}

\item{labels}{Names for on the pie-chart}

\item{edges}{?}

\item{radius}{What should be the radius of the whole pie-chart?}

\item{clockwise}{Do we want to go clockwise or counterclock-wise in the pie-chart?}

\item{init.angle}{At which angle should we start?}

\item{density}{?}

\item{angle}{?}

\item{col}{Which colors do you want to use within the pie-chart?}

\item{border}{Should there be a border? If so, what color should the border have?}

\item{lty}{Normally what type of line should be applied.}

\item{main}{What title should the pie-chart get? In our case, we do not use it.}

\item{...}{If you want to specify other things.}
}
\value{
Piece of the pie-chart.
}
\description{
Function to plot part of a piechart, which is required to 'build up' the
final piechart layer by layer.
}
