\name{vcov.mvl1lm}
\Rdversion{1.1}
\alias{vcov.mvl1lm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Variance-Covariance Matrix of an mvl1lm Object
}
\description{
Extracts the variance-covariance matrix of an mvl1lm Object.
}
\usage{
\method{vcov}{mvl1lm}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
an object of class \code{mvl1lm}.
}
  \item{\dots}{
needed for other methods.
}
}
\details{
For details see Chapter 13 of the MNM book. 
}

\references{
\cite{Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R}, Springer. } 
}
\author{Klaus Nordhausen, \email{klaus.nordhausen@uta.fi} }

\keyword{methods}
