% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeanAndSDAccuracy.R
\name{calculateMeanAndSDAccuracy}
\alias{calculateMeanAndSDAccuracy}
\title{Calculate performance of M&M on total train and test set}
\usage{
calculateMeanAndSDAccuracy(
  classColumn,
  higherClassColumn,
  minorityDir,
  majorityDir,
  metaDataRef,
  metaDataTest,
  throwOut,
  subtype = F,
  crossValidation = T,
  nModels,
  nSeeds,
  rounding,
  probabilityThreshold
)
}
\arguments{
\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}

\item{minorityDir}{Directory in which the minority model(s) are stored.}

\item{majorityDir}{Directory in which the majority model(s) are stored.}

\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the reference cohort.}

\item{metaDataTest}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the test set.}

\item{throwOut}{Are there samples you would like to remove from the test set due to poor data quality? If so, add their rownames here.}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?}

\item{crossValidation}{Specify whether the results are from the cross-validation setup or not.}

\item{nModels}{How many models were created for the majority voting system?}

\item{nSeeds}{How many seeds was the cross-validation setup run with?}

\item{rounding}{Do you want rounded numbers for the performance scores?}

\item{probabilityThreshold}{What is the probability score threshold you would like to use to call a classification 'confident'?}
}
\value{
Dataframe containing the average performance of tumor classifications within a certain frequency range (nCases).
Included are the averages for the percentage of correctly and incorrectly classified samples ($meanFractionCorrect and $meanFractionIncorrect),
correctly and incorrectly classified 'confident' samples ($meanFractionCorrectFiltered and $meanFractionIncorrectFiltered),
averaged precision values for the average of all tumor entities within the frequency range ($meanPrecision),
averaged recall values for the average of all tumor entities within the frequency range ($meanRecall),
and averaged F1 scores for the average of all tumor entities within the frequency range ($meanF1).
Please note that the precision, F1 and recall are calculated for the confident sample classifications only.
The total amount of samples within each frequency range ($meanSamples) is also specified.
}
\description{
Calculate performance of M&M on total train and test set
}
