% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeanAndSDInfluenceTumorHeterogeneity.R
\name{calculateMeanAndSDInfluenceTumorHeterogeneity}
\alias{calculateMeanAndSDInfluenceTumorHeterogeneity}
\title{Calculate the influence of tumor heterogeneity on performance M&M}
\usage{
calculateMeanAndSDInfluenceTumorHeterogeneity(
  minorityDir,
  majorityDir,
  metaDataRefWithStatus,
  metaDataTestWithStatus = NA,
  probabilityScoreTumor,
  probabilityScoreSubtype,
  metaDataFFPE = NA,
  predictionsFFPE,
  predictionsFFPESubtype
)
}
\arguments{
\item{minorityDir}{Directory in which the minority model(s) are stored.}

\item{majorityDir}{Directory in which the majority model(s) are stored.}

\item{metaDataRefWithStatus}{Metadata file containing the links between the samples and the tumor (sub)type diagnosis, and the tumor status (primary, metastasis, etc.).
Column names for the status are 'Status' and 'SystemicTreatment'.}

\item{metaDataTestWithStatus}{Metadata file containing the links between the samples and the tumor (sub)type diagnosis within the test set.}

\item{probabilityScoreTumor}{What is the probability score threshold you would like to use to call a classification 'confident' on the tumor type level?}

\item{probabilityScoreSubtype}{What is the probability score threshold you would like to use to call a classification 'confident' on the tumor subtype level?}

\item{metaDataFFPE}{Metadata file containing the links between the samples and the tumor (sub)type diagnosis for the FFPE samples.}

\item{predictionsFFPE}{Predictions for the FFPE samples on the tumor type level.}

\item{predictionsFFPESubtype}{Predictions for the FFPE samples on the tumor subtype level.}
}
\value{
Dataframe containing the results concerning the samples from different sources of tumor heterogeneity for supplied dataset.
}
\description{
Calculate the influence of tumor heterogeneity on performance M&M
}
