% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeanAndSDOverall.R
\name{calculateMeanAndSDOverall}
\alias{calculateMeanAndSDOverall}
\title{Calculate mean and standard deviation for whole dataset}
\usage{
calculateMeanAndSDOverall(
  classColumn,
  higherClassColumn,
  minorityDir,
  majorityDir,
  metaDataRef,
  metaDataTest,
  throwOut = NA,
  subtype = F,
  crossValidation = T,
  nModels,
  nSeeds,
  subset = F,
  probabilityThreshold
)
}
\arguments{
\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}

\item{minorityDir}{Directory in which the minority model(s) are stored.}

\item{majorityDir}{Directory in which the majority model(s) are stored.}

\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the reference cohort.}

\item{metaDataTest}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis within the test set.}

\item{throwOut}{Are there samples you would like to remove from the test set due to poor data quality? If so, add their rownames here.}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?}

\item{crossValidation}{Specify whether the results are from the cross-validation setup or not.}

\item{nModels}{How many models were created for the majority voting system?}

\item{nSeeds}{How many seeds was the cross-validation setup run with?}

\item{subset}{Do you want to only include a subset of the data for the performance measure calculations? If so, add their rownames here.}

\item{probabilityThreshold}{What is the threshold you would like to use to call a classification 'confident'?}
}
\value{
Dataframe to be used for
}
\description{
Calculate mean and standard deviation for whole dataset
}
