% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMulticlassROC.R
\name{calculateMulticlassROC}
\alias{calculateMulticlassROC}
\title{Title}
\usage{
calculateMulticlassROC(
  predictionsMMFinal,
  metaDataRef,
  classColumn,
  MMProbabilityList,
  filtering,
  probabilityScore
)
}
\arguments{
\item{predictionsMMFinal}{Dataframe containing the final classification ($predict) and original diagnosis label ($originalCall).}

\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis.}

\item{probabilityScore}{}
}
\description{
Title
}
