% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSeparateF1.R
\name{calculateSeparateF1}
\alias{calculateSeparateF1}
\title{Calculate performance measures for separate tumor (sub)types}
\usage{
calculateSeparateF1(
  minorityDir,
  majorityDir,
  subtype,
  metaDataTest = NA,
  probabilityThreshold
)
}
\arguments{
\item{minorityDir}{Directory in which the minority model(s) are stored.}

\item{majorityDir}{Directory in which the majority model(s) are stored.}

\item{subtype}{Do you want to obtain the classifications on the tumor subtype classification level?}

\item{metaDataTest}{Metadata file containing the links between the samples and the tumor (sub)type diagnosis within the test set}

\item{probabilityThreshold}{What is the threshold you would like to use to call a classification 'confident'?
If you do not want to filter the classifications, specify a score of 0.}
}
\value{
Dataframe containing the mean precision ($meanPrecision), F1-score ($meanF1), recall ($meanRecall)
and sensitivity per tumor (sub)type ($tumorType).
Results are stratified by the population frequency ($nCases).
Lastly, it's specified whether a cross-validation (Train) or test (Test) type was used.
}
\description{
Calculate performance measures for separate tumor (sub)types
}
