% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModelsMinority.R
\name{createModelsMinority}
\alias{createModelsMinority}
\title{Generate Minority classifier models}
\usage{
createModelsMinority(
  countDataRef,
  metaDataRef,
  patientColumn,
  meanExpression = 5,
  classColumn,
  nModels = 100,
  nANOVAgenes = 1000,
  maxSamplesPerType = 3,
  ntree = 500,
  howManyFeatures = 300,
  whichSeed = 1,
  outputDir = "./",
  proteinDir
)
}
\arguments{
\item{countDataRef}{Matrix containing the RNA-transcript per million data. Patients are in the columns,
different genes in the rows.}

\item{metaDataRef}{Reference cohort metadata file containing the links between the patients and the tumor (sub)type diagnosis.}

\item{patientColumn}{Column in the metadata file that contains the patient labels.}

\item{meanExpression}{Selection criterion for the genes,
specifying what the minimum mean expression of a gene should be for it to be included in the F-statistic analysis.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{nModels}{How many models should be created for the majority voting system?}

\item{nANOVAgenes}{How many genes should we select using the F-statistic of ANOVA?}

\item{maxSamplesPerType}{How many samples should we maximally use per tumor (sub)type?}

\item{ntree}{How many trees should we use during the weighted Random Forest (RF) procedure?}

\item{howManyFeatures}{How many features should we keep after determining the most important genes using the Random Forest Importance Score?}

\item{whichSeed}{For reproducibility, the seed can be specified with this parameter.}

\item{outputDir}{Directory in which you would like to store the R-object containing the results.}

\item{proteinDir}{In which directory can we find the file specifying the names of protein-coding genes within our dataset?}
}
\value{
R-object containing the generated RF-models ($modelList), the model for the ribodepletion correction ($riboModelList),
the features that were eventually used for the weighted RF within the different folds ($reducedFeaturesList),
the metadata file associated to the reference cohort ($metaData)
and the metadata for the performed run ($metaDataRun).

@import tidyverse dplyr magrittr randomForest caret glmnet
}
\description{
Setup to obtain the models for the Minority classifier.
}
