% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUMAPcohort.R
\name{createUMAPcohort}
\alias{createUMAPcohort}
\title{UMAP-transform reference cohort}
\usage{
createUMAPcohort(
  countDataRef,
  metaDataRef,
  classColumn,
  higherClassColumn,
  domainColumn,
  correctRibo = T,
  abbreviations = NA,
  proteinCodingGenes,
  whichSeed = 1
)
}
\arguments{
\item{countDataRef}{Matrix containing the RNA-transcript per million data. Patients are in the columns,
different genes in the rows.}

\item{metaDataRef}{Metadata file containing the links between the patients and
the tumor (sub)type diagnosis within the reference cohort.}

\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}

\item{domainColumn}{Column in the metadata file that contains the tumor domain labels.}

\item{correctRibo}{Do you want to perform a correction for the ribodepletion protocol on your dataset?
Default = FALSE, giving only the tumor type labels with the associated abbreviations.}

\item{abbreviations}{Dataframe containing the links between the tumor (sub)type,
the abbreviation required in the plot, and the domain.}

\item{whichSeed}{For reproducibility, the seed can be specified with this parameter.}

\item{proteinFile}{In which directory can we find the file specifying the names of protein-coding genes within our dataset?}
}
\value{
List containing the UMAP-transformed datapoints ($dataUMAP), and the ribodepletion correction model ($riboModelList).
}
\description{
Function to create the data required for plotting an UMAP of the reference cohort,
and paving the way for transforming new samples into the same sample space by
saving the transformation-model and ribo-depletion correction model.
}
\details{
Please note that the tumor domain, type and subtype all need to be specified within the metadata.
}
