% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUMAPcohort.R
\name{createUMAPcohort}
\alias{createUMAPcohort}
\title{UMAP-transform reference cohort}
\usage{
createUMAPcohort(
  countDataRef,
  metaDataRef,
  classColumn,
  higherClassColumn,
  domainColumn,
  sampleColumn,
  correctRibo = F,
  abbreviations = NA,
  noAbbreviations = F,
  proteinCodingGenes,
  whichSeed = 1
)
}
\arguments{
\item{countDataRef}{Matrix containing the RNA-transcript per million data. Samples are in the columns,
different genes in the rows.}

\item{metaDataRef}{Metadata file containing the links between the samples and
the tumor (sub)type diagnosis within the reference cohort.}

\item{classColumn}{Name of column in the metadata file that contains the tumor subtype labels.}

\item{higherClassColumn}{Name of column in the metadata file that contains the tumor type labels.}

\item{domainColumn}{Name of column in the metadata file that contains the tumor domain labels.}

\item{sampleColumn}{Name of column in the metadata file that contains the sample identifiers.}

\item{correctRibo}{Do you want to perform a correction for the ribodepletion protocol on your dataset? Default is TRUE.}

\item{abbreviations}{Optional. Dataframe containing the links between the tumor (sub)type,
the abbreviation required in the plot, and the domain.}

\item{proteinCodingGenes}{What are the names of the RNA-transcripts that stand for protein-coding genes within our dataset?
Please supply it as a vector. This is needed for ribo-depletion correction model.}

\item{whichSeed}{For reproducibility, the seed can be specified with this parameter.}
}
\value{
List containing the UMAP-transformed datapoints ($dataUMAP), the abbreviations used to name the UMAP-transformed datapoints, and the ribodepletion correction model ($riboModelList).
}
\description{
Function to create the data required for plotting an UMAP of the reference cohort
and ribo-depletion correction model.
}
\details{
Please note that the tumor domain, type and subtype all need to be specified within the metadata.
}
