% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccuraciesPerTumorTypeSize.R
\name{getAccuraciesPerTumorTypeSize}
\alias{getAccuraciesPerTumorTypeSize}
\title{Obtain accuracies per tumor (sub)type sample size}
\usage{
getAccuraciesPerTumorTypeSize(
  predictionsMM,
  metaDataRef,
  classColumn,
  rounding = T,
  probabilityThreshold
)
}
\arguments{
\item{predictionsMM}{Dataframe showing the top 3 predictions for the tumor (sub)type, together with their probability scores.}

\item{metaDataRef}{Metadata-file for the reference cohort containing the tumor (sub)type labels.}

\item{classColumn}{Which column within the metadata file contains the tumor (sub)type labels?}

\item{rounding}{Do you want rounded numbers for the performance scores?}

\item{probabilityThreshold}{What is the threshold you would like to use to call a classification 'confident'?}
}
\value{
Dataframe showing the performance scores for the different tumor frequencies (nCases).
Shown are the total amount of samples within the range ($nSamples),
how many are classified confidently ($nSamplesFiltered),
what is the fraction classified confidently ($fraction),
what is the fraction of correct and incorrect classifications within the highest scoring classification ($fractionCorrect and $fractionIncorrect),
the top 2  ($fractionCorrect2) an top 3 highest scoring classifications ($fractioncorrect3).
Furthermore, the fraction correct classifications within the confident classifications is denoted ($fractionCorrectFiltered),
together with the amount of errors remaining in here ($errorsFiltered).

For the rest, for the confident classifications the average precision ($Precision), F1 score ($F1) and recall ($Recall) for all tumor entities combined is determined.
}
\description{
Obtain accuracies per tumor (sub)type sample size
}
