% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfusionMatrixPlot.R
\name{getConfusionMatrixPlot}
\alias{getConfusionMatrixPlot}
\title{Get confusion matrix information}
\usage{
getConfusionMatrixPlot(
  minorityDir,
  majorityDir,
  abbreviations = NA,
  subtype,
  probabilityThreshold
)
}
\arguments{
\item{minorityDir}{Directory in which the Minority classifier model(s) are stored.}

\item{abbreviations}{Dataframe containing the links between the tumor subtype and their abbreviations ($abbreviationSubtype),
the tumor types and their abbreviations ($abbreviationTumorType), and the domain.}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?}

\item{probabilityThreshold}{Which probability score do you want to use as a cutoff for calling classifications confident or not?}

\item{majorityDirDirectory}{in which the Majority classifier model(s) are stored.}
}
\value{
List containing all tiles that contain classifications with their frequencies ($confusionPlotDF),
which tiles do not contain classifications ($nonAvailableTiles),
what were the abbreviations used as input for generating the  ($abbreviations),
what is the column in the abbreviations that contains the tumor subtype labels ($classColumn),
what is the column in the abbreviations that contains the tumor type labels ($higherClassColumn),
and lastly whether the tiles with classification frequencies were based on tumor type($subtype = F) or subtype level ($subtype = T).
}
\description{
This function is designed to extract which tiles of the confusion matrix
do contain reference-prediction information, and which do not.
IMPORTANT: The order of the tumor types within the plot will be equal to the order of the
tumors within the \emph{abbreviations} dataframe. Make sure you specify the
order of the tumor types well within the abbreviations, not only concerning the domains,
but also the subsequent tumor types order and tumor subtypes order.
}
