% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonAvailableTiles.R
\name{getNonAvailableTiles}
\alias{getNonAvailableTiles}
\title{Obtain empty tiles for confusion matrix}
\usage{
getNonAvailableTiles(
  predictionsMM,
  probabilityScore,
  abbreviations,
  classColumn
)
}
\arguments{
\item{predictionsMM}{Predictions for the tumor (sub)types with their associated probability scores.}

\item{probabilityScore}{Probability score that you would like to use as a threshold for calling samples
'classifiable' or 'non-classifiable'.}

\item{abbreviations}{Dataframe containing the abbreviations and full names of the tumor (sub)types, for the eventual confusion matrix plot labels.
The abbreviation should be stored in a column with header 'abbreviation'.}

\item{classColumn}{Specifies the name of the column of the abbreviations dataframe that contains the full tumor type label.}
}
\value{
A dataframe containing the empty tiles for a confusion matrix plot.
}
\description{
Get tiles from the confusion matrix that are never used for a prediction.
These tiles are important for the generation of the eventual confusion matrix.
}
