% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPieDF.R
\name{getPieDF}
\alias{getPieDF}
\title{Title}
\usage{
getPieDF(
  metaDataRef,
  abbreviationSubtype,
  abbreviationTumorType,
  highestClassColumn = "Disease_main_class",
  higherClassColumn,
  classColumn
)
}
\arguments{
\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis.}

\item{abbreviationSubtype}{Dataframe containing the link between the
tumor subtype label and tumor subtype abbreviation.}

\item{abbreviationTumorType}{Dataframe containing the link between the
tumor type label and tumor type abbreviation.}

\item{highestClassColumn}{Potentially not needed.}

\item{higherClassColumn}{Column in the metadata file that contains the tumor type labels.}

\item{classColumn}{Column in the metadata file that contains the tumor subtype labels.}
}
\value{
Dataframe containing ...
}
\description{
Title
}
