% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrincipalComponents.R
\name{getPrincipalComponents}
\alias{getPrincipalComponents}
\title{Get PCA transformation Majority Classifier}
\usage{
getPrincipalComponents(
  dataTrain,
  samplesTrainDefList,
  classColumn,
  nModels = 100,
  nFeatures = 2500,
  nComps = 100
)
}
\arguments{
\item{dataTrain}{Data to be used for training, containing the whole reference cohort.}

\item{samplesTrainDefList}{List of different training data subsets used for the majority voting system.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{nModels}{How many models should be created for the majority voting system?}

\item{nFeatures}{How many of the most variable genes within the dataset should we select for principal component analysis (PCA)?}

\item{nComps}{How many principal components will be selected after PCA?}
}
\value{
List containing both the derived PCA-transformation information ($prList)
and the information needed to scale new sample input data to center the features around 0 ($scaleFeaturesList).
}
\description{
Function to find the optimal principal component analysis (PCA) transformation of the reference cohort
so that a maximal variance within the data is explained. With the parameters, the number of variable features
and the eventual total number of principal components can be adjusted.
}
