% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeparateClassifierAccuracies.R
\name{getSeparateClassifierAccuracies}
\alias{getSeparateClassifierAccuracies}
\title{Get separate results from minority and majority classifier}
\usage{
getSeparateClassifierAccuracies(
  minority,
  majority,
  subtype = F,
  probabilityThresholdMajority,
  probabilityThresholdMinority
)
}
\arguments{
\item{minority}{R-object that contains the results from the Minority classifier}

\item{majority}{R-object that contains the results from the Majority classifier
This is important, as for the cross-validation setup there is a ground truth ($originalCall), while for new predictions there is not.}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level?}

\item{classColumn}{Column name within metadata-file that contains the cancer subtype-labels.}

\item{higherClassColumn}{Column name within metadata-file that contains the cancer type labels.}

\item{probabilityThreshold}{What is the probability score threshold that needs to be used to filter confident sample classifications with?}
}
\value{
Dataframe containing the total amount of samples ($nSamples), how many samples pass the the probability score threshold ($nSamplesFiltered),
what is the fraction non-classified samples ($notClassified),
what is the fraction of samples that pass the threshold ($fraction), what is the fraction of correct samples within the highest scoring classification ($fractionCorrect),
the top 2 scoring classification ($fractionCorrect2) and top 3 scoring classifications ($fractionCorrect3), what is the fraction of incorrect classifications ($fractionIncorrect),
how many of the confident classifications are correct ($fractionCorrectFiltered), and certain score metrics ($Precision, $F1 and $Recall),
all split up based on their population frequency within the reference cohohrt ($nCases) and the classifier used to obtain the metrics.
}
\description{
Get separate results from minority and majority classifier
}
