% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateMM.R
\name{integrateMM}
\alias{integrateMM}
\title{Integrate results Minority & Majority classifier}
\usage{
integrateMM(minority, majority, subtype, integrate = T)
}
\arguments{
\item{minority}{R-object that contains the results from the Minority classifier}

\item{majority}{R-object that contains the results from the Majority classifier}

\item{subtype}{Do you want to obtain the predictions on the tumor subtype classification level (TRUE) or
tumor type level (FALSE)?}

\item{integrate}{Do you want to integrate the results of the Minority and Majority classifiers into M&M,
or do you want to obtain the results from the separate classifiers?
If so, use \emph{subtype = T}. If you want to obtain tumor type classifications instead, use \emph{subtype = F}.}
}
\value{
If integrate = T: List containing the a dataframe with the final classifications (predictionsMMFinal),
and a list with the probability scores for all classification labels that were assigned to samples (MMProbabilityList).

predictionsMMFinal contains the top 3 final classification labels ($predict{2,3}) with their accompanying probability scores
($probability{1,2,3} and the original diagnosis label ($originalCall).

If integrate = F: List containing a dataframe with the final classifications separately for the Minority Classifier ($predictionsMinority)
and Majority Classifier ($predictionsMajority). This dataframe has the same setup as predictionsMMFinal.
}
\description{
Function to obtain final classifications for the M&M algorithm by combining the results of the Minority & Majority classifiers.
}
