\name{MNM-package}
\alias{MNM-package}
\alias{MNM}
\docType{package}
\title{
Multivariate Nonparametric Methods. An Approach Based on Spatial Signs and Ranks.}
\description{
The package provides multivariate tests, estimates and methods based on the identity score, spatial sign score and spatial rank score. The methods include one and c-sample problems, shape estimation and testing, 
linear regression and principal components.
}
\details{
\tabular{ll}{
Package: \tab MNM\cr
Type: \tab Package\cr
Version: \tab 0.95-3\cr
Date: \tab 2011-06-12\cr
License: \tab GPL (>= 2)\cr
}


The methods implemented here are mainly described in Oja (2010) and the package can be used to reproduce most of the examples in the book. The book will be refered to as the MNM book. 
}
\references{
\cite{Oja, H. (2010), \emph{Multivariate Nonparametric Methods with R. An Approach Based on Spatial Signs and Ranks}, Springer. } 
}
\author{
Klaus Nordhausen, Jyrki \enc{Möttönen}{Mottonen} and Hannu Oja

Maintainer: Klaus Nordhausen, \email{klaus.nordhausen@uta.fi} 

}

\keyword{package }
