% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPredictionsMajority.R
\name{newPredictionsMajority}
\alias{newPredictionsMajority}
\title{Predict tumor (sub)type for new samples with Majority Classifier}
\usage{
newPredictionsMajority(
  createdModelsMajority = createdModelsMajority,
  countDataRef,
  countDataNew,
  classColumn,
  nModels = 100,
  nComps = 100,
  maxNeighbours = 25,
  outputDir = "./",
  saveModel = T
)
}
\arguments{
\item{createdModelsMajority}{R-object containing the rotations and scalings
for each reference cohort subset($rotationsAndScalingList),
the model to correct for the ribodepletion efficacy ($riboModelList),
which samples were present in each subset ($samplesTrainDefList),
which genes were considered for transformation in the analysis ($nonZeroGenes),
the metadata file associated to the reference cohort ($metaData),
and metadata for the performed run ($metaDataRun).}

\item{countDataRef}{Matrix containing the RNA-transcript per million data for the reference cohort.
Patients are in the columns,
different genes in the rows.}

\item{countDataNew}{Matrix containing the RNA-transcript per million data for the new samples to be classified.
Patients are in the columns, different genes in the rows.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{nModels}{How many models should be created for the majority voting system?}

\item{nComps}{How many principal components will be selected after PCA?}

\item{maxNeighbours}{What is the maximum number of neigbours to be used for the weighted \emph{k}-nearest neighbor algorithm?}

\item{outputDir}{Directory in which you would like to store the R-object containing the results.}

\item{saveModel}{Do you want to save the resulting predictions in an R object?}
}
\value{
R-object containing the final classifications ($classifications) for the samples,
and the probabilities associated to the different classifications ($probability).
}
\description{
Function to generate predictions for new input samples from their
RNA-seq count data.
}
