% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPredictionsMinority.R
\name{newPredictionsMinority}
\alias{newPredictionsMinority}
\title{Predict tumor (sub)type for new samples with Minority Classifier}
\usage{
newPredictionsMinority(
  createdModelsMinority,
  countDataNew,
  outputDir = paste0("./", format(as.Date(Sys.Date(), "\%Y-\%m-\%d"), "\%Y_\%m_\%d")),
  saveModel = T,
  correctRibo = T,
  CPMcorrect = T,
  whichKimputation = 3
)
}
\arguments{
\item{createdModelsMinority}{R-object containing the generated RF-models ($modelList), the model for the ribodepletion correction ($riboModelList),
the features that were eventually used for the weighted RF within the different folds ($reducedFeaturesList),
the metadata file associated to the reference cohort ($metaDataRef)
and the metadata for the performed run ($metaDataRun).}

\item{countDataNew}{Matrix containing the RNA-transcript per million data for the new samples to be classified.
Samples are in the columns, different RNA-transcripts in the rows.}

\item{outputDir}{Directory in which you would like to store the R-object containing the results.}

\item{saveModel}{Do you want to save the results in an R object?}

\item{correctRibo}{Do you want to perform a correction for the ribodepletion protocol on your dataset? Default is TRUE.}

\item{CPMcorrect}{Do you want to make sure you have CPM counts? Put false in case when you supply a subset of RNA-transcripts at the start.}

\item{whichKimputation}{The number of neighbor datapoints that need to be used to calculate missing RNA-transcripts from (in case needed).}
}
\value{
R-object containing the final classifications ($classifications) for the samples,
the probabilities associated to the different classifications ($probability),
the metadata file associated to the reference cohort ($metaDataRef)
and the metadata for the performed run ($metaDataRun).
}
\description{
Predict tumor (sub)type for new samples with Minority Classifier
}
