% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainPredictionMajorityClassifier.R
\name{obtainPredictionMajorityClassifier}
\alias{obtainPredictionMajorityClassifier}
\title{Obtain predictions from generated majority models}
\usage{
obtainPredictionMajorityClassifier(
  rotationsAndScalingsList,
  dataTest,
  metaDataRef,
  testSamples,
  classColumn,
  maxNeighbors = 25,
  nModels = 100
)
}
\arguments{
\item{rotationsAndScalingsList}{List containing both the derived PCA-transformation information ($prList)
and the information needed to scale new sample input data to center the features around 0 ($scaleFeaturesList).}

\item{dataTest}{New samples for which the classifier should provide a classification label.}

\item{metaDataRef}{Metadata file containing the links between the samples and the tumor subtype diagnosis.}

\item{testSamples}{test sample IDs matching with samples from the test data.}

\item{classColumn}{Name of column in the metadata file that contains the tumor subtype labels.}

\item{maxNeighbors}{What is the maximum number of neighbors to be used for the weighted \emph{k}-nearest neighbor algorithm?}

\item{nModels}{How many models should be created for the Majority classifier?}
}
\value{
Dataframe containing the classifications for the \emph{nModels} different generated models, with the different folds
in the columns and the different samples to be predicted in the rows.
}
\description{
This function looks at the accompanying prediction for the test samples for the different generated Majority classifier models.
}
