% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainProbabilities.R
\name{obtainProbabilities}
\alias{obtainProbabilities}
\title{Obtain probabilities for classifier predictions}
\usage{
obtainProbabilities(classifierResults)
}
\arguments{
\item{classifierResults}{R-object containing the results from the minority or majority classifier.
The object should contain a list with the probabilities for each prediction.}
}
\value{
List of all the samples containing the probabilities for the different sample predictions from the \emph{nModels} generated models.
}
\description{
This function calculates the probability for each prediction, by dividing the number of models predicting a tumor (sub)type by the total number of models ( \emph{nModels}).
}
