% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainTrainData.R
\name{obtainTrainData}
\alias{obtainTrainData}
\title{Obtain training data subsets}
\usage{
obtainTrainData(
  metaDataRef,
  classColumn,
  maxSamplesPerType = 50,
  nModels = 100
)
}
\arguments{
\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis for the reference cohort.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{maxSamplesPerType}{How many samples should we maximally use per tumor (sub)type?}

\item{nModels}{How many models should be created for the majority voting system?}
}
\value{
A list with the specified number of different training data subsets from the reference cohort,
all with the specified maximum samples per tumor (sub)type.
}
\description{
This function is used to generate different training subsets for the majority voting system within the M&M setup.
}
