% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortTumorType.R
\name{plotCohortTumorType}
\alias{plotCohortTumorType}
\title{Plot UMAP tumor types}
\usage{
plotCohortTumorType(
  dataUMAP,
  domain,
  domainCol,
  classColumn,
  abbreviationTumorType,
  useLabels = T,
  useManualColors = F
)
}
\arguments{
\item{dataUMAP}{Dataframe containing the UMAP coordinates, the tumor type ($subclass),
the domain ($Domain) and the desired abbreviation for the tumor type ($abbreviation)}

\item{domain}{Which domain do you want to plot the tumor types for?}

\item{domainCol}{Which colors do you want to use for the tumor types?}

\item{abbreviationTumorType}{Abbreviations to be used in the plot for the tumor types.}

\item{useLabels}{Do you want to supply labels within the plot?}

\item{useManualColors}{Do you want to supply the colors to be used within the UMAP for the labels?}
}
\value{
ggplot with only the datapoints of the selected domain,
color coded by the tumor type.
}
\description{
Function to plot the UMAP coordinates for the different tumor type samples
belonging to one domain. All tumor types are colored differently to
distinguish them from one another.
}
