% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortTumorType.R
\name{plotCohortTumorType}
\alias{plotCohortTumorType}
\title{Plot UMAP tumor types}
\usage{
plotCohortTumorType(
  dataUMAPList,
  domain,
  tumorType = NA,
  plotColors = NA,
  subtype = F,
  useLabels = T
)
}
\arguments{
\item{domain}{Which domain do you want to plot the tumor types for?}

\item{tumorType}{If you show a tumor subtype instead of a tumor type (subtype = T),
please specify here which tumor types you would like to visualize (for example: "B-ALL")}

\item{plotColors}{Which colors do you want to use for the tumor types?}

\item{subtype}{Do you want to visualize the subtypes of one tumor type?
#' @param tumorType If subtype = T, which tumor type would you want to visualize?}

\item{useLabels}{Do you want to supply labels within the plot?}

\item{dataUMAP}{Dataframe containing the UMAP coordinates, the tumor type ($subclass),
the domain ($Domain) and the desired abbreviation for the tumor type ($abbreviation)}

\item{abbreviations}{Abbreviations to be used in the plot for the tumor (sub)types.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{useManualColors}{Do you want to supply the colors to be used within the UMAP for the labels?}
}
\value{
ggplot with only the datapoints of the selected domain,
color coded by the tumor type.
}
\description{
Function to plot the UMAP coordinates for the different tumor type samples
belonging to one domain. All tumor types are colored differently to
distinguish them from one another.
}
