% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortUMAP.R
\name{plotCohortUMAP}
\alias{plotCohortUMAP}
\title{Plot UMAP coordinates colored by tumor domain, tumor types or tumor subtypes}
\usage{
plotCohortUMAP(
  dataUMAPList,
  domain = NA,
  subtype = F,
  tumorType = NA,
  plotColors = NA,
  useLabels = T,
  labelSize = 4
)
}
\arguments{
\item{dataUMAPList}{List containing the UMAP coordinate data
and the desired abbreviation for the tumor type ($abbreviation), resulting from the function 'createUMAPcohort'}

\item{domain}{Which domain do you want to visualize the tumor types for (for example: domain = "Hemato")?
If not specified, visualization will be performed on the domain level.}

\item{subtype}{Do you want to visualize on the tumor subtype level (TRUE) or on the tumor type level (FALSE)?
Only possible when the domain is specified (e.g. domain = "Hemato"), otherwise visualization will
automatically take place on the domain level.}

\item{tumorType}{If domain is specified and subtype = TRUE, there's a possibility to plot the tumor subtypes for only one tumor type.
If that is desired, fill out the tumor type that you would like to visualize at 'tumorType' (for example: tumorType = "B-ALL")}

\item{plotColors}{Which colors do you want to use for the tumor types? If not specified, default ggplot2 colors will be used.}

\item{useLabels}{Do you want to supply labels within the plot?}

\item{labelSize}{What is the size that you would like to use for the labels on the UMAP? Default = 4.}
}
\value{
ggplot with only the datapoints of the selected domain,
color coded by the tumor type (subtype = F) or tumor subtype (subtype = T).
}
\description{
Function to plot the UMAP coordinates for the different tumor (sub)type samples
belonging to one domain or tumor type. Coloring and labeling can be performed based on the
sample's tumor domain, tumor type and tumor subtype labels.
}
