% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfusionMatrix.R
\name{plotConfusionMatrix}
\alias{plotConfusionMatrix}
\title{Create confusion matrix plot with the three domains.}
\usage{
plotConfusionMatrix(confusionPlotInfo)
}
\arguments{
\item{confusionPlotDF}{Dataframe specifying how often certain reference-prediction
combinations are present.}

\item{nonAvailableTiles}{A dataframe containing the empty tiles for a confusion matrix plot.}
}
\value{
ggplot object with the confusion matrix of the three domains.
}
\description{
This function can plot a confusion matrix both for the tumor type
and the tumor subtype level.
}
