% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfusionMatrixPerDomain.R
\name{plotConfusionMatrixPerDomain}
\alias{plotConfusionMatrixPerDomain}
\title{Plot tumor subtype confusion matrix}
\usage{
plotConfusionMatrixPerDomain(
  domain,
  confusionPlotInfo,
  domainColor = NA,
  colorTiles = "#012695"
)
}
\arguments{
\item{domain}{Which domain do you want to plot? This name should be available in the
domain column of the abbreviations.}

\item{confusionPlotInfo}{List resulting from running the function 'getConfusionMatrixPlot'.}

\item{domainColor}{Which colors should we use for each designated tumor type? If not specified, default ggplot colors will be used.}

\item{colorTiles}{Which domain color do we want to use?}
}
\value{
ggplot object containing the tumor subtype confusion matrix.
}
\description{
Function to plot the confusion matrix plot for a single domain,
showing the reference-prediction combinations for the tumor subtypes.
Using a color coding, it is shown within the plot which tumor subtypes belong to the same parent tumor type.
}
