% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleHeterogeneity.R
\name{plotSampleHeterogeneity}
\alias{plotSampleHeterogeneity}
\title{Plot results for data heterogeneity}
\usage{
plotSampleHeterogeneity(
  tumorHeterogeneityTrain,
  tumorHeterogeneityTest,
  subtype
)
}
\arguments{
\item{tumorHeterogeneityTrain}{Dataframe containing the results for different sources of heterogeneity within the training data}

\item{tumorHeterogeneityTest}{Dataframe containing the results for different sources of heterogeneity within the test data}

\item{subtype}{Do you want to perform the analyses on the tumor subtype classification level?}
}
\value{
Plot showing the accuracy, precision and recall for samples coming from different treatment statuses (treated, untreated)
and tumor statuses (primary, recurrence, metastasis)
}
\description{
Plot results for data heterogeneity
}
