% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeparateClassifierAccuracies.R
\name{plotSeparateClassifierAccuracies}
\alias{plotSeparateClassifierAccuracies}
\title{Plot the separate accuracy of the Minority classifier, Majority classifier and M&M classifier}
\usage{
plotSeparateClassifierAccuracies(separateClassifierAccuracies)
}
\arguments{
\item{separateClassifierAccuracies}{Dataframe obtained from the function 'combineSeparateClassifierAccuracies'.}
}
\value{
Plot showing the accuracy for different population frequencies separately for the Minority, Majority and M&M classifier.
}
\description{
Plot the separate accuracy of the Minority classifier, Majority classifier and M&M classifier
}
