% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeparateClassifierAccuracies.R
\name{plotSeparateClassifierAccuracies}
\alias{plotSeparateClassifierAccuracies}
\title{Title}
\usage{
plotSeparateClassifierAccuracies(
  minorityDir,
  majorityDir,
  classColumn,
  higherClassColumn,
  probabilityThreshold,
  probabilityThresholdMajority,
  probabilityThresholdMinority,
  subtype,
  nModels,
  nSeeds,
  returnPlot = T
)
}
\arguments{
\item{fractionsCorrectTotal}{}
}
\value{
plot with the separate classifier accuracies per frequency bin.
}
\description{
Title
}
