% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTop3Accuracy.R
\name{plotTop3Accuracy}
\alias{plotTop3Accuracy}
\title{Plot accuracy within the top 3 highest scoring classification labels M&M}
\usage{
plotTop3Accuracy(meanAndSDPlotTrain, meanAndSDPlotTest)
}
\arguments{
\item{meanAndSDPlotTrain}{Dataframe containing the average performance of tumor classifications within a certain frequency range (nCases) for the reference cohort.
Object obtained from running the function 'calculateMeanAndSDAccuracy'}

\item{meanAndSDPlotTest}{Dataframe containing the average performance of tumor classifications within a certain frequency range (nCases) for the test cohort.
Object obtained from running the function 'calculateMeanAndSDAccuracy'}
}
\value{
Plot showing the accuracy if not only the top scoring classification label, but also the second and third highest labels are taken into consideration.
Plot is segregated into the different class frequencies.
}
\description{
Plot accuracy within the top 3 highest scoring classification labels M&M
}
