% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precisionsForDomains.R
\name{precisionsForDomains}
\alias{precisionsForDomains}
\title{Title}
\usage{
precisionsForDomains(
  nSeeds,
  classColumn,
  higherClassColumn,
  domainColumn,
  crossValidation,
  minorityDir,
  majorityDir,
  nModels,
  subtype,
  throwOut,
  metaDataTest,
  probabilityThreshold,
  metaDataRef
)
}
\arguments{
\item{metaDataRef}{}
}
\description{
Title
}
