% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictTest.R
\name{predictTest}
\alias{predictTest}
\title{Classify test samples}
\usage{
predictTest(modelList, test.data)
}
\arguments{
\item{modelList}{List containing the models generated by the minority classifier.}

\item{test.data}{Test data to be classified. The dataset contains only the variables needed as input for the classifier.}
}
\value{
Dataframe containing the predictions of the tumor (sub)types for each sample within the minority classifier.
Each row is a test sample, with the separate columns containing the predictions from the separate models.
}
\description{
Function to classify new samples using the models generated with the reference cohort.
}
