% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceFeatures.R
\name{reduceFeatures}
\alias{reduceFeatures}
\title{Reduce features Minority Classifier with RF-importance}
\usage{
reduceFeatures(
  dataTrain,
  samplesTrainDefList,
  ntree = 500,
  nModels = 10,
  howManyFeatures = 300,
  nANOVAgenes
)
}
\arguments{
\item{dataTrain}{Data to be used for training, containing the whole reference cohort.}

\item{samplesTrainDefList}{List of different training data subsets used for the majority voting system.}

\item{ntree}{How many trees should we use during the weighted Random Forest (RF) procedure?}

\item{nModels}{How many models should be created for the majority voting system?}

\item{howManyFeatures}{How many features should we keep after determining
the most important genes using the Random Forest Importance Score?}
}
\value{
Names of the top most important genes present in the training data,
containing most information to distinguish the different tumor (sub)types.
}
\description{
This functions' purpose is to further reduce the number of features within the Minority Classifier.
Ten Random Forest (RF) models will be generated using the reference cohort training data,
from which the importance scores for the different genes will be extracted.
The \emph{howManyFeatures} genes with the highest importance scores will be selected for the final Minority Classifier training.
}
