% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFoldCrossValidationMinority.R
\name{tenFoldCrossValidationMinority}
\alias{tenFoldCrossValidationMinority}
\title{10x cross-validation minority classifier}
\usage{
tenFoldCrossValidationMinority(
  countDataRef,
  metaDataRef,
  meanExpression = 5,
  classColumn,
  patientColumn,
  nModels = 100,
  nANOVAgenes = 1000,
  maxSamplesPerType = 3,
  ntree = 500,
  howManyFeatures = 300,
  whichSeed = 1,
  outputDir = "./",
  proteinDir =
    "~/surfdrive/Shared/Kemmeren group/Research_Projects/RNA_classification_FW/data/input/"
)
}
\arguments{
\item{countDataRef}{Matrix containing the RNA-transcript per million data. Patients are in the columns,
different genes in the rows.}

\item{metaDataRef}{Metadata file containing the links between the patients and the tumor (sub)type diagnosis.}

\item{meanExpression}{Selection criterion for the genes,
specifying what the minimum mean expression of a gene should be for it to be included in the F-statistic analysis.}

\item{classColumn}{Column in the metadata file that contains the tumor (sub)type labels.}

\item{patientColumn}{Column in the metadata file that contains the patient labels.}

\item{nModels}{How many models should be created for the majority voting system?}

\item{nANOVAgenes}{How many genes should we select using the F-statistic of ANOVA?}

\item{maxSamplesPerType}{How many samples should we maximally use per tumor (sub)type?}

\item{ntree}{How many trees should we use during the weighted Random Forest (RF) procedure?}

\item{howManyFeatures}{How many features should we keep after determining the most important genes using the Random Forest Importance Score?}

\item{whichSeed}{For reproducibility, the seed can be specified with this parameter.}

\item{outputDir}{Directory in which you would like to store the R-object containing the results.}

\item{proteinDir}{In which directory can we find the file specifying the names of protein-coding genes within our dataset?}
}
\value{
R-object containing the predictions ($classifications), classifications errors ($wrongClassifications),
the probabilities for each classification ($probabilityList), the metadata file associated to the reference cohort ($metaData),
the features that were eventually used for the weighted RF within the different folds ($reducedFeaturesList)
and the metadata for the performed run ($metaDataRun).
}
\description{
Setup to automatically run the 10x stratified cross-validation minority classifier.
}
